% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adj_matrix_from_nei_list}
\alias{adj_matrix_from_nei_list}
\title{Return adjacency matrix from list of neighbours}
\usage{
adj_matrix_from_nei_list(nei_list)
}
\arguments{
\item{nei_list}{a list of vectors, where each vector contains the indices of the adjacent neighbours.}
}
\value{
a sparse adjacency matrix (with zero on the diagonal)
}
\description{
Creates a sparse adjacency matrix from list of vertex neighbours.
}
\examples{
nei_list <- list(c(2,3),1,1)
adj.matrix <- adj_matrix_from_nei_list(nei_list)
}
\author{
Andrew Zammit Mangion
}

