% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{AIRS_05_2003}
\alias{AIRS_05_2003}
\title{AIRS data for May 2003}
\format{A data frame with 223662 rows and 8 variables:
\describe{
  \item{year}{year of retrieval}
  \item{month}{month of retrieval}
  \item{day}{day of retrieval}
  \item{lat}{latitude coordinate of retrieval}
  \item{lon}{longitude coordinate of retrieval}
  \item{solzen}{solar zenith angle}
  \item{co2avgret}{CO2 mole fraction in ppm}
  \item{co2std}{standard deviation of CO2 reading in ppm}
}}
\usage{
AIRS_05_2003
}
\description{
Mid-troposheric CO2 measurements from the Atmospheric InfraRed Sounder (AIRS). The data represent measurements between 60 degrees S and 90 degrees North at roughly 1:30 pm local time on 1 May through 16 May 2003
}
\references{
Chahine, M., Pwaagepeagano, T., Aumann, H., Atlas, R., Barnet, C., Blaisdell, J., Chen, L., Divakarla, M., Fetzer, E., Goldberg, M., Gautier, C., Granger, S., Hannon, S., Irion, F. W., Kakar, R., Kalnay, E., Lambrigtsen, B. H., Lee, S.-Y., Marshall, J. L., McMillian, W. W., McMillin, L., Olsen, E. T., Revercomb, H., Rosenkranz, P., Smith, W. L., Staelin, D., Strow, L. L., Susskind, J., Tobin, D., Wolf, W., Zhou, L. (2006). AIRS: Improving weather forecasting and providing new data on greenhouse gases. Bulletin of the American Meteorological Society 87, 911–26.
}
\keyword{datasets}

