% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[<--methods}
\alias{[<-,Basis,ANY,ANY,ANY-method}
\alias{[<-,Basis,ANY,ANY-method}
\alias{[<-,GMRF,ANY,ANY,ANY-method}
\alias{[<-,GMRF,ANY,ANY-method}
\alias{[<-,GMRF_basis,ANY,ANY,ANY-method}
\alias{[<-,GMRF_basis,ANY,ANY-method}
\alias{[<-,Obs,ANY,ANY,ANY-method}
\alias{[<-,Obs,ANY,ANY-method}
\alias{[<--methods}
\title{Assignment methods}
\usage{
\S4method{[}{GMRF_basis,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[}{GMRF,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[}{Obs,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[}{Basis,ANY,ANY,ANY}(x, i, j) <- value
}
\description{
Methods for \code{"[<-"}, i.e., extraction or subsetting of elements in the data frame of the MVST object
}
\examples{
data(surf_fe)
Mesh <- initFEbasis(p=surf_fe$p, t = surf_fe$t, M = surf_fe$M, K = surf_fe$K)
Mesh["z"] <- sin(Mesh["x"]/1000)*cos(Mesh["y"]/1000)
}

