% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attribute_polygon}
\alias{attribute_polygon}
\title{Polygon-attribution}
\usage{
attribute_polygon(df, shape_df)
}
\arguments{
\item{df}{data frame to which the polygon id is attributed. Must contain fields \code{x} and \code{y}.}

\item{shape_df}{the polygon-containing data frame. Must contain fields \code{x}, \code{y} and \code{id}. Multiple polygons/land-masses can be treated simultaneously through \code{id}.}
}
\value{
a vector of length \code{nrow(df)} indicating the id in which each point in \code{df} lies.
}
\description{
Takes a data frame \code{df} with fields \code{x} and \code{y} and a shape in data frame format, \code{shape_table}, and determines whether, and in which, polygon each point in \code{df} lies.
}
\examples{
df <- data.frame(x=c(0.2,1),y=c(0.2,1))
shape_df <- data.frame(x=c(0,0,0.5,0.5),
y = c(0,0.5,0.5,0),
id = 1)
df$id <- attribute_polygon(df,shape_df)
}

