% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/geometryfns.R
\docType{methods}
\name{auto_BAU}
\alias{auto-BAU,plane-method}
\alias{auto_BAU}
\alias{auto_BAU,plane-method}
\alias{auto_BAU,sphere-method}
\title{Automatic BAU generation}
\usage{
auto_BAU(manifold, cellsize, res, type, data, ...)

\S4method{auto_BAU}{plane}(manifold, cellsize, res, type, data, ...)

\S4method{auto_BAU}{sphere}(manifold, cellsize, res, type, data, ...)
}
\description{
Automatic BAU generation
}

