% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{auto_BAUs}
\alias{auto_BAUs}
\title{Automatic BAU generation}
\usage{
auto_BAUs(manifold, res = 2, cellsize = rep(1, dimensions(manifold)),
  type = "hex", data = NULL, convex = -0.05)
}
\description{
Automatic BAU generation
}

