% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/basisfns.R
\docType{data}
\name{Basis_as_data.frame}
\alias{Basis_as_data.frame}
\alias{TensorP_Basis_as_data.frame}
\alias{data.frame<-}
\alias{$,Basis-method}
\alias{$<-,Basis-method}
\alias{data.frame<-,Basis-method}
\alias{data.frame_Basis,Basis-method}
\alias{data.frame<-,TensorP_Basis-method}
\alias{data.frame_Basis,Basis-method}
\alias{as.data.frame.Basis}
\alias{as.data.frame.TensorP_Basis}
\title{Basis-function data frame object}
\format{An object of class \code{NULL} of length 0.}
\usage{
Basis_as_data.frame

TensorP_Basis_as_data.frame

data.frame(x) <- value

\\S4method{$}{Basis}(x, name)

\\S4method{$}{Basis}(x, name) <- value

\\S4method{data.frame}{Basis}(x) <- value

\\S4method{data.frame}{TensorP_Basis}(x) <- value

\\method{as.data.frame}{Basis}(x, ...)

\\method{as.data.frame}{TensorP_Basis}(x, ...)
}
\arguments{
\item{x}{the obect of class \code{Basis} we are assigning the new data to or retrieving data from}

\item{value}{the new data being assigned to the Basis object}

\item{name}{the field name to which values will be retrieved or assigned inside the Basis' data frame}

\item{...}{unused}
}
\description{
Tools for retrieving and manipulating the data frame within the Basis objects. Use the assignment \code{data.frame()<-} with care; no checks are made to make sure the data frame conforms with the object. Only use if you know what you're doing
}
\examples{
G <- local_basis()
df <- data.frame(G)
print(df$res)
df$res <- 2
data.frame(G) <- df
}
\keyword{datasets}
