% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Basis-class}
\alias{Basis-class}
\title{Parent class for basis functions}
\description{
A basis function contains three slots, the first is a list of parameters (pars), the second is the number (n) of basis functions and the third is a list
of functions (fn). The pars field usually contains parameters which appear in fn.
}
\details{

}
\keyword{Basis}
\keyword{functions}

