% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{bisquareBasis-class}
\alias{bisquareBasis-class}
\title{Bisquare radial basis functions basis}
\description{
bisquareBasis inherits from the virtual class Basis. A bisquare basis functions is initialsied using the function \code{initbisquarebasis}.
}
\details{

}
\keyword{Basis}
\keyword{GRBF}
\keyword{functions,}

