% Generated by roxygen2 (4.0.2): do not edit by hand
\name{block_list}
\alias{block_list}
\title{List of blocks}
\usage{
block_list(l = NULL)
}
\arguments{
\item{l}{a list of objects of class \code{block}.}
}
\value{
Object of class \code{block_list} (which inherits from class \code{list})
}
\description{
This function initialises an object of class \code{block_list} which is simply a list in which each object is restricted to be of class \code{block}. A \code{block}
is also either of class \code{Obs} or class \code{process}.
}
\examples{
\dontrun{
require(Matrix)
data(icesat)
data(surf_fe)

## First create observation object
icesat_obs <- Obs(df=icesat,
                 abs_lim = 5,
                 avr_method = "median",
                 box_size=100,
                 name="icesat")

## Now create GMRF defined over some FE basis
Mesh <- initFEbasis(p=surf_fe$p,
                    t=surf_fe$t,
                    M=surf_fe$M,
                    K=surf_fe$K)

mu <- matrix(0,nrow(Mesh),1)
Q <- sparseMatrix(i=1:nrow(surf_fe$p), j = 1:nrow(surf_fe$p), x = 1)

my_GMRF <- GMRF(mu = mu, Q = Q,name="SURF",t_axis = 0:6)
SURF <-GMRF_basis(G = my_GMRF, Basis = Mesh)

v <- block_list(list(O = icesat_obs, G = SURF))
}
}
\keyword{Process}
\keyword{blocks}
\keyword{blocks,}
\keyword{observation}

