% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{block_list-class}
\alias{block_list-class}
\title{List of blocks}
\description{
The list of blocks is simply a collection of blocks, which are either latent process or observations. The block list is combined with a link list in order to
compose a graph.
}
\details{

}

