% Generated by roxygen2 (4.0.2): do not edit by hand
\name{border_from_grid_points}
\alias{border_from_grid_points}
\title{Extract border}
\usage{
border_from_grid_points(df, delta)
}
\arguments{
\item{df}{a data frame with fields \code{x} and \code{y}.}

\item{delta}{the grid spacing.}
}
\value{
a subset of \code{df} containing the border points.
}
\description{
The border is extracted from shapes supplied in x-y format
}
\details{
The algorithm works by finding how many neighbours a grid point has using \code{fields::fields.rdist.near}. A grid point with less than 4 neighbours is assumed to be a border point. For this algorithm to work, it is important that the shape is `full' and not contain any missing values.
}
\examples{
grid <- expand.grid(x = c(1:10), y = c(1:10))
names(grid) = c("x","y")
border_points <- border_from_grid_points(df=grid,delta=1)
}
\author{
Andrew Zammit Mangion
}

