% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linalgfns.R
\name{cholsolve}
\alias{cholsolve}
\title{Solve the equation Qx = y}
\usage{
cholsolve(Q, y, perm = F, cholQ = matrix(1, 0, 0), cholQp = matrix(1, 0,
  0), P = NA)
}
\arguments{
\item{Q}{matrix (sparse or dense), the Cholesky factor of which needs to be found}

\item{y}{matrix with the same number of rows as Q}

\item{perm}{if F no permutation is carried out, if T permuted Cholesky factors are used}

\item{cholQ}{the Cholesky factor of Q (if known already)}

\item{cholQp}{the permuted Cholesky factor of Q (if known already)}

\item{P}{the pivot matrix (if known already)}
}
\value{
x solution to Qx = y
}
\description{
This function is similar to \code{solve(Q,y)} but with the added benefit that it allows for permuted matrices. This function does the job in order to minimise
user error when attempting to re-permute the matrices prior or after solving. The user also has an option for the permuted Cholesky factorisation of Q to be carried out
internally.
}
\examples{
require(Matrix)
Q = sparseMatrix(i=c(1,1,2,2),j=c(1,2,1,2),x=c(0.1,0.2,0.2,1))
y = matrix(c(1,2),2,1)
cholsolve(Q,y)
}
\references{
Havard Rue and Leonhard Held (2005). Gaussian Markov Random Fields: Theory and Applications. Chapman & Hall/CRC Press
}
\keyword{Cholesky}
\keyword{factor,}
\keyword{linear}
\keyword{solve}

