% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/basisfns.R
\docType{methods}
\name{concat}
\alias{concat}
\alias{concat,Basis-method}
\title{Concatenation}
\usage{
concat(...)

\S4method{concat}{Basis}(...)
}
\arguments{
\item{...}{a series of \code{MVST} objects}
}
\description{
Concatenates MVST objects of the same class together. This is primarily used to join up \code{GMRF_basis} blocks and \code{Obs} blocks together.
}

