% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ConstBasis-class}
\alias{ConstBasis-class}
\title{Constant functions basis}
\description{
ConstBasis inherits from the virtual class Basis. A set of constant basis functions is initialsied using the function \code{initConstbasis}.
}
\details{

}
\keyword{#'}
\keyword{Basis}
\keyword{functions}

