% Generated by roxygen2 (5.0.1): do not edit by hand
\name{dggrid_gen_to_df}
\alias{dggrid_gen_to_df}
\title{Convert DGGRID gen. file to data frame}
\usage{
dggrid_gen_to_df(filename, res)
}
\arguments{
\item{filename}{name of DGGRID output file of type .gen}

\item{res}{resolution of the DGGRID (this number is attached to the resulting data frame)}
}
\description{
Convert discrete global grid (DGGRID) .gen file to data frame
}
\details{
DGGRID produces .gen files with three columns, \code{id}, \code{lon} and \code{lat}. Polygons each have a unique identifier and following each polygon there is a line "END" that needs to be filtered out. This function takes the .gen file and converts it into a data frame with fields \code{lon, lat, id} and \code{res}.
}

