% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBmisc.R
\name{.concat_percentiles_to_df}
\alias{.concat_percentiles_to_df}
\title{Computation and concatenation of percentiles to a dataframe.}
\usage{
.concat_percentiles_to_df(data, X, name, interval_type, percentiles)
}
\arguments{
\item{data}{The dataframe we will append percentiles to; the number of rows (if \code{X} is a matrix) or the length (if \code{X} is a list) of \code{X} must equal the number of rows in \code{data}}

\item{X}{X a matrix (wherein rows correspond to prediction location, columns are samples)  or a list (wherein each element contains a vector of samples at a given prediction location) of Monte Carlo samples}

\item{name}{The name of the quantity of interest. The names of the percentile columns are "name_percentile". In \code{FRK} it will be Y, mu, prob, or Z}

\item{interval_type}{string indicating whether a \code{"central"} or \code{"HPD"} interval is desired}

\item{percentiles}{a vector of scalars in [0, 100] specifying the desired percentiles (applicable only if \code{interval_type = "central"}); if \code{percentiles = NULL}, no percentiles are computed}
}
\value{
The dataframe \code{data} with appended percentiles
}
\description{
Computes the percentiles or HPD interval bounds at each prediction location and appends 
the result to \code{data}. Note that we use percentiles rather than quantiles
because we including a "dot" (corresponding to the decimal place) in the 
dataframe column name may cause issues.
}
