% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.cov_tap}
\alias{.cov_tap}
\title{Covariance tapering based on distances.}
\usage{
.cov_tap(D_matrices, taper = 8)
}
\arguments{
\item{D_matrices}{A list of distance matrices corresponding to each resolution.}

\item{taper}{The strength of the taper (either a vector or a single number).}
}
\value{
A list containing:
\describe{
  \item{alpha}{A vector of taper parameters.}
  \item{nnz}{A vector containing the number of non-zeros in each block of the 
  tapered prior covariance matrix, K_tap.}
}
}
\description{
Computes the covariance tapering parameters \eqn{\alpha} (which are dependent
on resolution) and the number of non-zeros in each block of the tapered
covariance matrix K_tap.
}
\details{
\code{taper} determines how strong the covariance tapering is; the ith taper parameter
\eqn{\alpha_i} is equal to \code{taper[i]} * \code{minDist[i]}, where
\code{minDist[i]} is the minimum distance between basis functions at the
\code{i}th resolution.
}
\seealso{
\code{\link{.K_matrix}}
}
