% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{.diagnostic_stats_FRK}
\alias{.diagnostic_stats_FRK}
\title{Summary of diagnostic functions for the FRK model}
\usage{
.diagnostic_stats_FRK(M, newdata, MC = NULL, true_value, out_sample = TRUE)
}
\arguments{
\item{M}{SRE object.}

\item{newdata}{Dataframe returned by \code{predict}.}

\item{MC}{List of Monte Carlo samples returned by \code{predict}. This argument
is only used to compute the continuous ranked probability score; set \code{MC = NULL}
if this is not desired.}

\item{true_value}{Named list, where the names indicate the quantities of interest.
Must contain at least one element named 'Y', 'mu', or 'prob'.}

\item{out_sample}{Are we interested in out of sample locations? If \code{out_sample = TRUE}, 
only the unobserved BAUs are used for computing the diagnostics.}
}
\value{
A data.frame containing diagnostics for each quantity of interest.
}
\description{
This function provides a summary of several diagnostic functions used to 
assess performance of the model. The diagnostics are root mean squared prediction error, 
the interval score (assuming a 90% interval), coverage, and, if a Monte Carlo sample is supplied, 
the continuous ranked probability score. The function assumes the same format that 
the \code{predict()} function returns when \code{method = 'TMB'}.
}
