% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpredict.R
\name{.FRKTMB_pred}
\alias{.FRKTMB_pred}
\title{Prediction stage of non-Gaussian FRK.}
\usage{
.FRKTMB_pred(
  M,
  newdata,
  CP,
  predict_BAUs,
  pred_time,
  type,
  n_MC,
  obs_fs,
  k,
  percentiles,
  kriging
)
}
\arguments{
\item{M}{An object of class SRE.}

\item{newdata}{object of class \code{SpatialPoylgons} indicating the regions over which prediction will be carried out. The BAUs are used if this option is not specified}

\item{CP}{Polygon prediction matrix}

\item{predict_BAUs}{Logical indicating whether or not we are predicting over the BAUs}

\item{pred_time}{vector of time indices at which prediction will be carried out. All time points are used if this option is not specified}

\item{type}{A character string (possibly vector) indicating the quantities for which predictions and prediction uncertainty is desired. If \code{"link"} is in \code{type}, the latent \eqn{Y} process is included; If \code{"mean"} is in \code{type}, the conditional mean \eqn{\mu} is included (and the probability parameter if applicable); If \code{"response"} is in \code{type}, the response variable \eqn{Z} is included. Note that any combination of these character strings can be provided. For example, if \code{type = c("link", "response")}, then predictions of the latent \eqn{Y} process and the response variable \eqn{Z} are provided}

\item{n_MC}{A postive integer indicating the number of MC samples at each location}

\item{obs_fs}{Logical indicating whether the fine-scale variation is included in the latent Y process. If \code{obs_fs = FALSE} (the default), then the fine-scale variation term \eqn{\xi} is included in the latent \eqn{Y} process. If \code{obs_fs = TRUE}, then the the fine-scale variation terms \eqn{\xi} are removed from the latent Y process; \emph{however}, they are re-introduced for computation of the conditonal mean \eqn{\mu} and response variable \eqn{Z}}

\item{k}{vector of size parameters at each BAU (applicable only for binomial and negative-binomial data)}

\item{percentiles}{a vector of scalars in [0, 100] specifying the desired percentiles; if \code{percentiles = NULL}, no percentiles are computed}

\item{kriging}{A string indicating the kind of kriging (\code{"simple"} or \code{"universal"})}
}
\value{
A list object containing:
\describe{
  \item{newdata}{An object of class \code{newdata}, with predictions and prediction uncertainty at each prediction location of the latent \eqn{Y} process, the conditional mean of the data \eqn{\mu}, the probability of success parameter \eqn{\pi} (if applicable), and the response variable \eqn{Z}}
  \item{MC}{A list with each element being an \code{N * n_MC} matrix of Monte Carlo samples of the quantities specified by \code{type} (some combination of \eqn{Y}, \eqn{\mu}, \eqn{p} (if applicable), and \eqn{Z}) at each prediction location}
}
Note that for all link functions other than the log- and identity-link functions, the predictions and prediction uncertainty of \eqn{\mu} contained in \code{newdata} are computed using the Monte Carlo samples contained in \code{MC}.
When the log- or identity-link functions are used, the expectation and variance of the \eqn{\mu} may be computed exactly.
}
\description{
Prediction stage of non-Gaussian FRK.
}
