% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBmisc.R
\name{.intervalWidth}
\alias{.intervalWidth}
\title{Prediction interval width.}
\usage{
.intervalWidth(X, l = 0.025, u = 0.975)
}
\arguments{
\item{X}{A matrix of Monte Carlo samples (rows are observation, columns are 
samples).}

\item{l}{The lower probability.}

\item{u}{The upper probability.}
}
\value{
The prediction interval width at each location.
}
\description{
This function is used to compute prediction interval width
given \code{X}, a matrix of Monte Carlo samples (wherein rows correpsond to locations, columns samples)
}
