% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.K_alpha_matrix}
\alias{.K_alpha_matrix}
\title{\eqn{K_\alpha} matrix, the taper matrix.}
\usage{
.K_alpha_matrix(D_matrices, alpha)
}
\arguments{
\item{D_matrices}{A list of distance matrices corresponding to each resolution.}

\item{alpha}{The taper parameters (which vary based on resolution).}
}
\value{
A taper matrix, which is block-diagonal and of class \code{dgCmatrix}.
}
\description{
Constructs \eqn{K_\alpha}, the taper matrix, using the Spherical taper.
}
\details{
Denoting by \eqn{d(s,s^*)} the distance between two spatial locations,
the Spherical taper is given by:
\deqn{C_\alpha(s, s^*) = (1-\frac{d(s,s^*)}{\alpha})^2_{+}  (1+\frac{d(s,s^*)}{2\alpha}), }
where \eqn{x_+ = max(x, 0)}.
}
