% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.K_beta_matrix}
\alias{.K_beta_matrix}
\title{\eqn{K_\beta} matrix, the taper matrix.}
\usage{
.K_beta_matrix(D_matrices, beta)
}
\arguments{
\item{D_matrices}{A list of distance matrices corresponding to each resolution.}

\item{beta}{The taper parameters (which vary based on resolution).}
}
\value{
A taper matrix, which is block-diagonal and of class \code{dgCmatrix}.
}
\description{
Constructs \eqn{K_\beta}, the taper matrix, using the Spherical taper.
}
\details{
Denoting by \eqn{d(s,s^*)} the distance between two spatial locations,
the Spherical taper is given by:
\deqn{C_\beta(s, s^*) = (1-\frac{d(s,s^*)}{\beta})^2_{+}  (1+\frac{d(s,s^*)}{2\beta}), }
where \eqn{x_+ = max(x, 0)}.
}
