% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.K_matrix}
\alias{.K_matrix}
\title{K matrix, the un-tapered prior covariance matrix.}
\usage{
.K_matrix(D_matrices, sigma2, tau)
}
\arguments{
\item{D_matrices}{A list of distance matrices corresponding to each resolution
(see \code{dist_matrices}).}

\item{sigma2}{The variance components: a vector with length equal to the number
of resolutions i.e. \code{length(D_matrices)}.}

\item{tau}{The correlation components: a vector with length equal to the number
of resolutions i.e. \code{length(D_matrices)}.}
}
\value{
The block-diagonal (class \code{dgCmatrix}) prior covariance matrix, K.
}
\description{
Construct the prior covariance matrix of the random effects (NOT-TAPERED).
}
\seealso{
\code{\link{dist_matrices}}
}
