% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.K_tap_matrix}
\alias{.K_tap_matrix}
\title{K_tap, the tapered prior covariance matrix.}
\usage{
.K_tap_matrix(D_matrices, beta, sigma2, tau)
}
\arguments{
\item{D_matrices}{A list of distance matrices corresponding to each resolution.}

\item{beta}{The taper parameters (which vary based on resolution).}

\item{sigma2}{The variance components: a vector with length equal to the number
of resolutions i.e. \code{length(D_matrices)}.}

\item{tau}{The correlation components: a vector with length equal to the number
of resolutions i.e. \code{length(D_matrices)}.}
}
\value{
The \emph{tapered} prior covariance matrix, which is
block-diagonal and of class \code{dgCmatrix}.
}
\description{
Construct K_tap, the \emph{tapered} prior covariance matrix, using the Spherical taper.
}
