% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.link_fn}
\alias{.link_fn}
\title{Link and inverse-link function generator.}
\usage{
.link_fn(kind, link, response)
}
\arguments{
\item{kind}{A character string indicating which kind of link function is desired. 
Valid values are 
\describe{
  \item{\code{"Y_to_mu"}}{Provides the function \eqn{\psi} such that \eqn{\mu = \psi(Y)}.}
  \item{\code{"mu_to_Y"}}{Provides the function \eqn{g} such that \eqn{g(\mu) = Y}.}
  \item{\code{"Y_to_prob"}}{Provides the function \eqn{\zeta} such that \eqn{p = \zeta(Y)}.}
  \item{\code{"prob_to_Y"}}{Provides the function \eqn{h} such that \eqn{h(p) = Y}.}
  \item{\code{"prob_to_mu"}}{Provides the function \eqn{\chi} such that \eqn{\mu = \chi(p)}.}
  \item{\code{"mu_to_prob"}}{Provides the function \eqn{f} such that \eqn{f(\mu) = p}.}
}
Note that the latter four values are relevant only to the binomial and negative-binomial distributions with logit, probit, or cloglog link functions.}

\item{link}{A character string indicating the assumed link function. \emph{Not} required if \code{kind} is \code{"prob_to_mu"} or \code{"mu_to_prob"}.}

\item{response}{A character string indicating the assumed response distribution. \emph{Only} required if \code{kind} is \code{"prob_to_mu"} or \code{"mu_to_prob"}.}
}
\value{
A function.
}
\description{
Create the link function \eqn{g}  and inverse-link function \eqn{\psi},
which link the conditional mean of the data \eqn{\mu} to the latent
geostatistical process \eqn{Y} such that \eqn{g(\mu) = Y} and
\eqn{\mu = \psi(Y)}, respectively.
}
\details{
For families lacking a "known constant" parameter,
\code{.link_fn} generates the functions linking
the conditional mean of the data \eqn{\mu} to the latent
geostatistical process \eqn{Y}. However, for families with a "known constant"
parameter (such as the "binomial" and "negative-binomial"), \code{.link_fn} generate the functions linking
the probability parameter \eqn{p} to the latent
geostatistical process \eqn{Y}, and then the parameter \eqn{p} to the conditional mean \eqn{\mu}.
}
