% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{.plot_all}
\alias{.plot_all}
\title{Plot non-Gaussian data and predictions.}
\usage{
.plot_all(newdata = NULL, zdf = NULL, response = NULL, x = "x", y = "y")
}
\arguments{
\item{newdata}{A \code{dataframe} containing the sspatial coordinates (named "x" and "y"), and the predictions and uncertainty quantification.}

\item{zdf}{A \code{dataframe} containing spatial coordinates (named "x" and "y") and the value of the observations (named "z").}

\item{response}{A character indicating the assumed response distribution.}

\item{x}{A \code{string} indicating the name of the column containing x-locations.}

\item{y}{A \code{string} indicating the name of the column containing the y-locations.}
}
\value{
A list of ggplot() objects.
}
\description{
Plot non-Gaussian data and predictions.
}
