% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBplot.R
\name{.plot_data}
\alias{.plot_data}
\title{Plot of spatial data.}
\usage{
.plot_data(zdf, response, point_size = 1, lim = range(zdf[["z"]]))
}
\arguments{
\item{zdf}{A \code{dataframe} containing spatial coordinates (named "x" and "y") and the value of the response variable (named "z").}

\item{response}{A character indicating the assumed response distribution.}

\item{point_size}{Size of plotted points.}

\item{lim}{Controls the limits of the colour scale.}
}
\value{
A \code{ggplot} of the data.
}
\description{
Plot of spatial data.
}
\seealso{
\code{\link{.plot_map}}, \code{\link{.plot_all}}
}
