% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{.plot_map}
\alias{.plot_map}
\title{Plot of spatial process.}
\usage{
.plot_map(
  df,
  col,
  x = "x",
  y = "y",
  lim = range(df[[col]]),
  diverging = FALSE,
  low = "blue",
  mid = "white",
  high = "red",
  midpoint = mean(df[[col]]),
  uncertaintyMap = FALSE
)
}
\arguments{
\item{df}{A \code{dataframe} containing spatial coordinates (named "x" and "y")
and the value of the process (whose name is specified by the \code{col} argument).}

\item{col}{A \code{string} indicating the name of the column containing the process values.}

\item{x}{A \code{string} indicating the name of the column containing x-locations.}

\item{y}{A \code{string} indicating the name of the column containing the y-locations.}

\item{lim}{Controls the limits of the colour scale.}

\item{diverging}{\code{logical}, indicating whether a diverging palette should be used.}

\item{low}{Low colour (only applicable if \code{diverging == TRUE}).}

\item{mid}{Mid colour (only applicable if \code{diverging == TRUE}).}

\item{high}{High colour (only applicable if \code{diverging == TRUE}).}

\item{midpoint}{Point at which colour \code{mid} is assigned. This is useful for
uncertainty ratio maps (in which case a good midpoint is 1) and probability maps
(in which case a good midpoint is 0.5).}

\item{uncertaintyMap}{Logical indicating whether to use uncertainty colour scale.}
}
\value{
A \code{ggplot} object of the spatial process.
}
\description{
Plot of spatial process.
}
\seealso{
\code{\link{.plot_data}}, \code{\link{.plot_all}}
}
