% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBmisc.R
\name{.prediction_interval}
\alias{.prediction_interval}
\title{Prediction intervals.}
\usage{
.prediction_interval(X, interval_type, percentiles, cred_mass)
}
\arguments{
\item{X}{a matrix (wherein rows correspond to prediction locations and columns are samples) of Monte Carlo samples}

\item{interval_type}{string indicating whether a \code{"central"} or \code{"HPD"} interval is desired}

\item{percentiles}{a vector of scalars in [0, 100] specifying the desired percentiles; if \code{percentiles = NULL}, no percentiles are computed}

\item{cred_mass}{a scalar [0, 1] specifying the mass within the credible interval; if \code{cred_mass = NULL}, no HPD credible interval is computed}
}
\value{
The prediction interval at each location (or width at each location if \code{width = TRUE})
}
\description{
Given a matrix of Monte Carlo samples \code{X} (wherein rows correspond to 
locations, columns correspond to samples), this function computes symmetric 
or HPD prediction intervals, and returns the prediction interval lower and 
upper bound
}
