% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBmisc.R
\name{.rowVars}
\alias{.rowVars}
\title{Variance by rows.}
\usage{
.rowVars(X, ...)
}
\arguments{
\item{X}{An array like object (with a dim-attribute).}
}
\value{
A vector containing the row-wise variances of the matrix \code{X}.
}
\description{
Computes the row-wise variance of a matrix.
}
