% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.sparse_Q}
\alias{.sparse_Q}
\title{Sparse precision matrix.}
\usage{
.sparse_Q(A, kappa, rho)
}
\arguments{
\item{A}{A "neighbour" matrix with element (i, j), for i not equal to j,
equal to 1 if basis functions i and j are neighbours, and 0 otherwise,
diagonal elements indicating the number of neighbours for that basis function.}

\item{kappa}{Quantity to add to the diagonal elements. This must be positive if Q is to be positive definite.}

\item{rho}{Quantity to multiply matrix by. This must be positive if Q is to be positive definite.}
}
\value{
A sparse precision matrix of class \code{dgCMatrix}.
}
\description{
Creates a sparse precision matrix \eqn{Q} with off diagonal elements equal
to -1 if the basis functions are neighbours, and zero otherwise.
The diagonal elements are equal to the number of neighbours for that basis
function, plus some amount given by \code{kappa}.
}
\seealso{
\code{\link{.neighbour_matrix}}, \code{\link{.sparse_Q_block_diag}}
}
