% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.sparse_Q_block_diag}
\alias{.sparse_Q_block_diag}
\title{Block-diagonal sparse precision matrix.}
\usage{
.sparse_Q_block_diag(df, kappa, rho, order = 1, diag_neighbours = FALSE)
}
\arguments{
\item{df}{A dataframe containing the spatial coordinates (named "loc1" and "loc2") and a blocking column (named "res").}

\item{kappa}{Quantity to add to the diagonal elements. This must be positive if Q is to be positive definite.}

\item{rho}{Quantity to multiply matrix by. This must be positive if Q is to be positive definite.}
}
\value{
A list containing the sparse block-diagonal precision matrix (Q) of class "dgCMatrix", and the number of non-zero elements (nnz) at each resolution.
}
\description{
Creates a block-diagonal sparse precision matrix, where the blocks are created
using \code{sparse_Q}.
}
\seealso{
\code{\link{.sparse_Q}}, \code{\link{.neighbour_matrix}}
}
