% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.TMB_prep}
\alias{.TMB_prep}
\title{TMB data and parameter preparation.}
\usage{
.TMB_prep(M, known_sigma2fs)
}
\arguments{
\item{M}{An object of class \code{SRE}.}

\item{known_sigma2fs}{known value of the fine-scale variance. If \code{NULL} (the default), the fine-scale variance \eqn{\sigma^2_\xi} is estimated as usual. If \code{known_sigma2fs} is not \code{NULL}, the fine-scale variance is fixed to the supplied value; this may be a scalar, or vector of length equal to the number of spatial BAUs (if fs_by_spatial_BAU = TRUE)}
}
\value{
A list object containing:
\describe{
  \item{data}{The data.}
  \item{parameters}{The initialised parameters/fixed-effects/random-effects.}
}
}
\description{
Prepares data and parameter/fixed-effects/random-effects for input to \code{TMB}.
}
