% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBpredict.R
\name{.Y_var}
\alias{.Y_var}
\title{Posterior variance of the latent Y process.}
\usage{
.Y_var(M, Q_L, obsidx)
}
\arguments{
\item{M}{An object of class SRE.}

\item{Q_L}{A list containing the Cholesky factor of the permuted precision matrix (stored as \texttt{Q$Qpermchol}) and the associated permutationmatrix (stored as \texttt{Q_L$P}).}

\item{obsidx}{Vector containing the observed locations.}
}
\value{
A vector of the posterior variance of Y at every BAU.
}
\description{
Computes the variance of the latent process \eqn{Y} at every BAU. Note that MSPE(E(Y|Z), Y) is approximated by var(Y|Z).
}
\details{
To compute the prediction uncertainty of Y we require the joint
covariance matrix of the random effects \eqn{(\eta', \xi_O')'}. \code{TMB}
provides an approximation of the joint \emph{precision} matrix of
\eqn{(\eta', \xi_O')'}, which we must invert to obtain the approximate
covariance matrix. However, due to the potentially very large number of
random effects (the number of observations \eqn{m} is not restricted),
in practice we compute the sparse-inverse-subset of the joint precision matrix,
\emph{not} the true joint covariance matrix.

Note that as we are using E(\eqn{Y|Z}) to predict \eqn{Y}, the posterior variance acts as an approximation of the mean-squared prediction error (see pg. 72 of Honours thesis).
}
