% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FEBasis-class}
\alias{FEBasis-class}
\title{Finite element basis}
\description{
FEBasis inherits from the virtual class Basis. A fintie element basis is initialsied using the function \code{initFEbasis}.
}
\details{

}
\keyword{Basis}
\keyword{functions}

