% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_hull}
\alias{find_hull}
\title{Find convex hull from a set of points}
\usage{
find_hull(points)
}
\arguments{
\item{points}{a data frame with fields \code{x} and \code{y}}
}
\value{
subset of \code{points} corresponding to the convex hull
}
\description{
Returns the convex hull from a set of points with fields \code{x} and \code{y}
}
\examples{
N = 99
points <- data.frame(x = runif(n=N), y = runif(n=N),id=1:N)
hull <- find_hull(points)
plot(points$x,points$y)
lines(hull$x,hull$y,col="red")
}
\author{
Andrew Zammit Mangion
}

