% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Findalphabeta_gamma}
\alias{Findalphabeta_gamma}
\title{Quantile difference when tuning Gamma distribution hyper-parameters}
\usage{
Findalphabeta_gamma(pars, p5, p95)
}
\arguments{
\item{pars}{the shape and rate parameters (in that order) of the Gamma distribution}

\item{p5}{the 5 percentile of the desired error distribution}

\item{p95}{the 95 percentils of the desired error distribution}
}
\value{
A numeric which is a measure of the discrepancy between the Gamma distribution over the precision and the 5/95 percentiles of the error.
}
\description{
This function returns the normalised difference between specified and required 5 and 95 percentiles of the Gamma distribution. The percentiles represent those of the required error, i.e. the root of the variance. For example, if the error of an observation is believed to be between 1 and 5, with 90% confidence, then we can see how distant a Gamma distribution over the precision with shape and rate parameters in \code{pars} differs from this belief by calling \code{Findalphabeta_gamma(pars,1,5)}. This function can then be passed to an optimisation routine to find better values for \code{pars}.
}
\examples{
require(actuar)
# Find the Gamma distribution over the precision corresponding to my prior belief of the error (1/sqrt(precision)) lying between p5=2, p95=10
initpars <- c(5,0.1)
hyp_pars <- optim(par=initpars,Findalphabeta_gamma, p5=1, p95=5)

# Now simulate from a Gamma with these parameters and verify quantiles
X <- rgamma(shape = hyp_pars$par[1], rate = hyp_pars$par[2],n=10000)
print( quantile(1/sqrt(X),c(0.05,0.95)))
}
\keyword{Gamma}
\keyword{distribution,}
\keyword{elicitation}
\keyword{prior}

