% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Findalphabeta_invgamma}
\alias{Findalphabeta_invgamma}
\title{Quantile difference when tuning inverse Gamma distribution hyper-parameters}
\usage{
Findalphabeta_invgamma(pars, p5, p95)
}
\arguments{
\item{pars}{the shape and scale parameters (in that order) of the inverse-Gamma distribution}

\item{p5}{the 5 percentile of the desired error distribution}

\item{p95}{the 95 percentils of the desired error distribution}
}
\value{
A numeric which is a measure of the discrepancy between the inverse-Gamma distribution and the 5/95 percentiles.
}
\description{
This function returns the normalised difference between specified and required 5 and 95 percentiles of the inverse gamma distribution over the variance. The percentiles represent those of the required error, i.e. the root of the variance. For example, if the error of an observation is believed to be between 2 and 10, with 90% confidence, then we can see how distant an inverse gamma distribution with shape and scale parameters in \code{pars} differs from this belief by calling \code{Findalphabeta_invgamma(pars,2,10)}. This function can then be passed to an optimisation routine to find better values for \code{pars}.
}
\examples{
require(actuar)
# Find the an inverse-gamma distribution of the variance parameter corresponding to my prior belief of the error (sqrt(variance)) lying between p5=2, p95=10
initpars <- c(5,0.1)
hyp_pars <- optim(par=initpars,Findalphabeta_invgamma, p5=2, p95=10)

# Now simulate from an inverse Gamma with these parameters and verify quantiles
X <- rinvgamma(shape = hyp_pars$par[1], scale = hyp_pars$par[2],n=10000)
print( quantile( sqrt(X),c(0.05,0.95)))
}
\keyword{Gamma}
\keyword{Inverse}
\keyword{distribution,}
\keyword{elicitation}
\keyword{prior}

