% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRK.R
\docType{package}
\name{FRK-package}
\alias{FRK-package}
\title{Fixed Rank Kriging}
\description{
Fixed Rank Kriging is a tool for spatial/spatio-temporal modelling and prediction with large datasets. The approach, discussed in Cressie and Johannesson (2008), decomposes the field, and hence the covariance function, using a fixed set of n basis functions, where dimension n is typically much smaller than the number of data points (or polygons) m. The method naturally allows for non-stationary, anisotropic covariance functions and the use of observations with varying support (with known error variance). The dimension-reduced field is a key building block of the Spatial Random Effects (SRE) model, upon which this package is based. The package FRK provides helper functions to model, fit, and predict using an SRE with relative ease. Reference: Cressie, N. and Johannesson, G. (2008) <DOI:10.1111/j.1467-9868.2007.00633.x>.
}
