% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getC}
\alias{getC}
\alias{getC,Graph_2nodes-method}
\alias{getC,linkGO-method}
\title{Get incidence matrix}
\usage{
getC(.Object)

\S4method{getC}{Graph_2nodes}(.Object)

\S4method{getC}{linkGO}(.Object)
}
\arguments{
\item{.Object}{object of class \code{Graph_2nodes} or class \code{link}.}
}
\value{
Object of class \code{Matrix}.
}
\description{
Retrieves the incidence matrix \code{C} from the observation model \eqn{z = Cx + e}.
}
\examples{
\dontrun{
require(Matrix)
data(icesat)
data(surf_fe)

## First create observation object
icesat_obs <- Obs(df=icesat,
                 abs_lim = 5,
                 avr_method = "median",
                 box_size=100,
                 name="icesat")

## Now create GMRF defined over some FE basis
Mesh <- initFEbasis(p=surf_fe$p,
                    t=surf_fe$t,
                    M=surf_fe$M,
                    K=surf_fe$K)

mu <- matrix(0,nrow(Mesh),1)
Q <- sparseMatrix(i=1:nrow(surf_fe$p), j = 1:nrow(surf_fe$p), x = 1)

my_GMRF <- GMRF(mu = mu, Q = Q,name="SURF",t_axis = 0:6)
SURF <-GMRF_basis(G = my_GMRF, Basis = Mesh)

L1 <- link(SURF,icesat_obs)
C_L1 <- getC(L1)
e <- link_list(list(L1))
v <- block_list(list(O = icesat_obs, G = SURF))
G <- new("Graph",e=e,v=v)
G_reduced <- compress(G)
C_Graph <- getC(G_reduced)
}
}
\keyword{incidence}
\keyword{matrix,}
\keyword{model}
\keyword{observation}

