% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getDf}
\alias{getDf}
\alias{getDf,FEBasis-method}
\alias{getDf,GMRF-method}
\alias{getDf,GMRF_basis-method}
\alias{getDf,Graph-method}
\alias{getDf,Obs-method}
\title{Get data frame}
\usage{
getDf(.Object)

\S4method{getDf}{Graph}(.Object)

\S4method{getDf}{FEBasis}(.Object)

\S4method{getDf}{GMRF}(.Object)

\S4method{getDf}{GMRF_basis}(.Object)

\S4method{getDf}{Obs}(.Object)
}
\arguments{
\item{.Object}{object from which data frame will be extracted}
}
\description{
Returns data frame from MVST block object. Can be used on objects of class \code{FEBasis,GMRF} and \code{Obs}. If used on an object of class \code{Graph}, the observation data frame is returned.
}
\examples{
data(icesat)
icesat_obs <- Obs(df=icesat)
df <- getDf(icesat_obs)
}

