% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getMean}
\alias{getMean}
\alias{getMean,GMRF-method}
\alias{getMean,GMRF_basis-method}
\title{Get mean vector}
\usage{
getMean(.Object)

\S4method{getMean}{GMRF}(.Object)

\S4method{getMean}{GMRF_basis}(.Object)
}
\arguments{
\item{.Object}{object from which mean vector will be extracted}
}
\description{
Returns the mean from MVST object.
}
\examples{
my_RW <- GMRF_RW(n=10, order=1, precinc =2, name="my_first_RW")
mu <- getMean(my_RW)
}

