% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getPrecision}
\alias{getPrecision}
\alias{getPrecision,GMRF-method}
\alias{getPrecision,GMRF_basis-method}
\alias{getPrecision,Obs-method}
\title{Get precision matrix}
\usage{
getPrecision(.Object)

\S4method{getPrecision}{Obs}(.Object)

\S4method{getPrecision}{GMRF}(.Object)

\S4method{getPrecision}{GMRF_basis}(.Object)
}
\arguments{
\item{.Object}{object from which precision matrix will be extracted}
}
\description{
Returns precision matrix from MVST object.
}
\examples{
my_RW <- GMRF_RW(n=10, order=1, precinc =2, name="my_first_RW")
Q <- getPrecision(my_RW)
}

