% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GMRF_basis}
\alias{GMRF_basis}
\title{GMRF function over basis}
\usage{
GMRF_basis(G = new("GMRF"), Basis = new("Basis", pars = list(vars =
  data.frame(x = c(1, 2)))))
}
\arguments{
\item{G}{an object of class \code{GMRF}}

\item{Basis}{an object of class \code{Basis}}
}
\value{
Object of class \code{GMRF_basis} (which inherits from class \code{process} and is thus also a process block)
}
\description{
This function initialises an object of class \code{GMRF_basis} which defines a GMRF over a set of basis functions.
}
\examples{
G <- GMRF_RW(n=9)
Basis <-  initGRBFbasis(x = c(0,1), y = c(0,1), std=0.1,nx=9,ny=1)
print(GMRF_basis(G,Basis))
}
\keyword{GMRF,}
\keyword{basis}
\keyword{functions}

