% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GMRF_RW-class}
\alias{GMRF_RW-class}
\title{GMRF_RW}
\description{
A random walk  which inherits from class \code{GMRF}. The primary difference is that this class is constructed using
a first-order auto-regressice structure. All random walks are intrinsic GMRFs.
}
\details{

}

