% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Graph}
\alias{Graph}
\title{Graph}
\usage{
Graph(e = new("link_list"), v = new("block_list"))
}
\arguments{
\item{e}{an object of class \code{link_list} containing all the edges in the graph.}

\item{v}{an object of class \code{block_list} containing all the blocks in the graph.}
}
\value{
Object of class \code{Graph}
}
\description{
A graph is a collection of links (collected using \code{link_list}) and blocks (collected using \code{block_list}).
}
\examples{
\dontrun{
require(Matrix)
data(icesat)
data(surf_fe)

## First create observation object
icesat_obs <- Obs(df=icesat,
                 abs_lim = 5,
                 avr_method = "median",
                 box_size=100,
                 name="icesat")

## Now create GMRF defined over some FE basis
Mesh <- initFEbasis(p=surf_fe$p,
                    t=surf_fe$t,
                    M=surf_fe$M,
                    K=surf_fe$K)

mu <- matrix(0,nrow(Mesh),1)
Q <- sparseMatrix(i=1:nrow(surf_fe$p), j = 1:nrow(surf_fe$p), x = 1)

my_GMRF <- GMRF(mu = mu, Q = Q,name="SURF",t_axis = 0:6)
SURF <-GMRF_basis(G = my_GMRF, Basis = Mesh)

L1 <- link(SURF,icesat_obs)
e <- link_list(list(L1))
v <- block_list(list(O = icesat_obs, G = SURF))
G <- new("Graph",e=e,v=v)
}
}
\keyword{Process}
\keyword{blocks,}
\keyword{graph}
\keyword{observation}

