% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Graph_2nodes-class}
\alias{Graph_2nodes-class}
\title{Graph_2nodes}
\description{
This class is a special case of \code{Graph}, where the link list and the block list only contain one node each. In general, a graph with many edges
can be compressed to a 2-node graph (one block latent process and one block observation set) using \code{compress}.
}

