% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{icesat}
\alias{icesat}
\title{ICESAT data.}
\format{A data frame with 10000 rows and 6 variables}
\description{
A sample of the Ice, Cloud and Elevation dataset, where each point represents an average trend in m/yr over a 20km x 20km grid box. The data has the following fields
}
\details{
\itemize{
  \item t. year of observation (0--6 where 0 is 2003, and 6 is 2009)
  \item x. x coordinate in km (under polar stereographic projection)
  \item y. y coordinate in km (under polar stereographic projection)
  \item std. std standard error of trend
  \item z. trend
  \item obs_name. the name of the observation
}
}
\keyword{datasets}

