% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init_radial_basis}
\alias{init_radial_basis}
\title{Initialise a GRBF basis}
\usage{
init_radial_basis(manifold, loc, scale, type = "Gaussian")
}
\description{
This function initialises an object of class \code{GRBFBasis} which defines a set of radial basis functions at pre-specified locations
}

