% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initbisquarebasis}
\alias{initbisquarebasis}
\title{Initialise a GRBF basis}
\usage{
initbisquarebasis(manifold, c, R)
}
\description{
This function initialises an object of class \code{GRBFBasis} which defines a set of radial basis functions at pre-specified locations
}

