% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initGRBFbasis}
\alias{initGRBFbasis}
\title{Initialise a GRBF basis}
\usage{
initGRBFbasis(x, y, std, nx, ny)
}
\arguments{
\item{x}{x-coordinate of GRBF centroid}

\item{y}{y-coordinate of GRBF centroid}

\item{std}{the 'length' (in terms of sigma) of the GRBF}

\item{nx}{the number of columns of the GRBF array}

\item{ny}{the number of rows of the GRBF array}
}
\value{
Object of class \code{GRBFBasis}
}
\description{
This function initialises an object of class \code{GRBFBasis} which defines a set of radial basis functions at pre-specified locations in 2-D
}
\examples{
Basis <-  initGRBFbasis(x = c(0,1), y = c(0,1), std=0.1,nx=9,ny=1)
}
\keyword{GRBF,}
\keyword{basis}
\keyword{functions}

