% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\docType{methods}
\name{initialize,manifold-method}
\alias{initialize,manifold-method}
\title{manifold}
\usage{
\S4method{initialize}{manifold}(.Object)
}
\description{
Buffer manifold initialisation. This function should not be called directly.
}

