% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kappa_from_l}
\alias{kappa_from_l}
\title{Get scale from practical range}
\usage{
kappa_from_l(l, nu)
}
\arguments{
\item{l}{the practical range.}

\item{nu}{the smoothness parameter}
}
\value{
numeric
}
\description{
Takes a practical range \code{l} and the smoothness of the Matern kernel \eqn{\nu} and returns the scale parameter \eqn{\kappa}.
}
\examples{
l = 100
nu = 1
kappa <- kappa_from_l(l,nu)
}
\seealso{
The inverse \code{\link{l_from_kappa}}
}
\keyword{Matern,}
\keyword{practical}
\keyword{range}

