% Generated by roxygen2 (4.0.2): do not edit by hand
\name{link}
\alias{link}
\title{link}
\usage{
link(Obj1, Obj2, Cmat = NULL, mul_factor = NULL, mulfun = NULL,
  muldata = NULL, n_grid = NULL, mask = NULL, md5_wrapper = NULL)
}
\arguments{
\item{Obj1}{object of class \code{process}.}

\item{Obj2}{object of class \code{Obs}.}

\item{Cmat}{the matrix mapping the process to the observation. Needs to be specified if no basis function set is assoicated with the process.}

\item{mul_factor}{a constant with which to scale the process, i.e. \eqn{y = mul_factor * C * x}}

\item{mulfun}{as above, but a possibly spatially varying amplification of the process of the observation. This, for example, could be a spatially-varying
density function when converting height to mass.}

\item{n_grid}{the number of grid points to use when integrating over the process with an observation of a large footprint. Defaults to a 400 x 400 grid.}

\item{mask}{the label of a process attribute. The observation is assumed to only be influenced by the process where this attribute is set to 1.}
}
\value{
Object of class \code{linkGO}, a link between a GMRF and an observation.
}
\description{
This function takes am object of class \code{process} and an object of class \code{Obs} and creates a link between the two.
If the \code{process} is of class \code{GMRF_basis}, then an incidence matrix is constructed which maps the process to the observation. If
there is no basis set associated with the GMRF, then the incidence matrix needs to be specified manually.
}
\examples{
\dontrun{
require(Matrix)
data(icesat)
data(surf_fe)

## First create observation object
icesat_obs <- Obs(df=icesat,
                 abs_lim = 5,
                 avr_method = "median",
                 box_size=100,
                 name="icesat")

## Now create GMRF defined over some FE basis
Mesh <- initFEbasis(p=surf_fe$p,
                    t=surf_fe$t,
                    M=surf_fe$M,
                    K=surf_fe$K)

mu <- matrix(0,nrow(Mesh),1)
Q <- sparseMatrix(i=1:nrow(surf_fe$p), j = 1:nrow(surf_fe$p), x = 1)

my_GMRF <- GMRF(mu = mu, Q = Q,name="SURF",t_axis = 0:6)
SURF <-GMRF_basis(G = my_GMRF, Basis = Mesh)

L1 <- link(SURF,icesat_obs)
}
}
\keyword{incidence}
\keyword{link,}
\keyword{matrix}

