% Generated by roxygen2 (4.0.2): do not edit by hand
\name{link_list}
\alias{link_list}
\title{List of links}
\usage{
link_list(l = NULL)
}
\arguments{
\item{l}{a list of objects of class \code{link}.}
}
\value{
Object of class \code{link_list} (which inherits from class \code{list})
}
\description{
This function initialises an object of class \code{link_list} which is simply a list in which each object is restricted to be of class \code{link}.
}
\examples{
\dontrun{
require(Matrix)
data(icesat)
data(surf_fe)

## First create observation object
icesat_obs <- Obs(df=icesat,
                 abs_lim = 5,
                 avr_method = "median",
                 box_size=100,
                 name="icesat")

## Now create GMRF defined over some FE basis
Mesh <- initFEbasis(p=surf_fe$p,
                    t=surf_fe$t,
                    M=surf_fe$M,
                    K=surf_fe$K)

mu <- matrix(0,nrow(Mesh),1)
Q <- sparseMatrix(i=1:nrow(surf_fe$p), j = 1:nrow(surf_fe$p), x = 1)

my_GMRF <- GMRF(mu = mu, Q = Q,name="SURF",t_axis = 0:6)
SURF <-GMRF_basis(G = my_GMRF, Basis = Mesh)

L1 <- link(SURF,icesat_obs)
e <- link_list(list(L1))
}
}
\keyword{Observations,}
\keyword{incidence}
\keyword{matrix}

