% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{link_list-class}
\alias{link_list-class}
\title{List of links}
\description{
The list of links is simply a collection of links, each with its own incidence matrix. The link list is combined with a block list in order to
compose a graph.
}
\details{

}

