% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{link-class}
\alias{link-class}
\title{Link between latent process and observation}
\description{
The prime use of the \code{link} class is to generate an incidence matrix between observational data and the process.
This class is a virtual class, and is a parent for the specific \code{linkGO} class which specifies a link between a process block
and an observational block
}
\details{

}

