% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Load_data}
\alias{Load_data}
\title{Load data}
\usage{
Load_data(thinning = 1, tif_grid = 750, convert_m_to_km = T, ...)
}
\arguments{
\item{thinning}{determines a thinning rate for the shapefiles. This is useful when shapefiles, particularly coastlines, are very high in resolution.}

\item{tif_grid}{the native resolution of the tif file, if present, in m}

\item{convert_m_to_km}{if \code{TRUE}, shapefiles and tif file units are converted to km}

\item{...}{paths to data sets which will be loaded. These need to be assigned variable names}
}
\value{
list of data frames
}
\description{
Loads data simultaneously from multiple files. For data files, it is assumed that the columns have a header. The function understands filenames with extensions 'shp', 'txt', 'dat', and 'tif'.
}
\examples{
df <- data.frame(x=c(1,2,3),y=c(2,2,2))
write.table(df,file="~/temp.dat")
data <- Load_data("~/temp.dat")
file.remove("~/temp.dat")
}

