% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Load_meshes}
\alias{Load_meshes}
\title{Load meshes}
\usage{
Load_meshes(paths)
}
\arguments{
\item{paths}{a list of path names to the relevant folders}
}
\value{
a list of objects of class \code{FEBasis}. Please refer to the vignette for and example.
}
\description{
Imports meshes from folders containing the files \code{p.csv, t.csv, M.csv} and \code{K.csv} which are csv files containing the vertices, triangulations, mass matrix and stiffness matrix details respectively. The matrices should be stored in the three-column format \code{i,j,k}.
}

