% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{loglik}
\alias{loglik}
\alias{loglik,SRE-method}
\title{(Deprecated) Retrieve log-likelihood}
\usage{
loglik(object)

\S4method{loglik}{SRE}(object)
}
\arguments{
\item{object}{object of class \code{SRE}}
}
\description{
This function is deprecated; please use \code{logLik}
}
