% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lscale_from_Matern}
\alias{lscale_from_Matern}
\title{Find length scales by fitting a Matern}
\usage{
lscale_from_Matern(data, rho = 100, nu = 3/2, var = 1,
  theta = seq(-0.99, 0.99, 0.33))
}
\arguments{
\item{data}{a data frame with fields \code{x}, \code{y} and an output variable \code{z}. The time coordinate \code{t} is optional.}

\item{rho}{an array of spatial practical ranges to consider.}

\item{nu}{an array of smoothness parameters to consider.}

\item{var}{an array of marginal variances to consider}

\item{theta}{an array of first-order auto-regressive parameters to consider in the model \eqn{x_{t+1} = \theta x_{t} + e_{t}}.}
}
\description{
Takes a spatial or spatio-temporal data set and returns spatial and temporal length scales by finding the maximum likelihood
on a pre-specified grid. If the data is spatio-temporal a separable covariance structure is assumed.
}
\examples{
var_true <- 1
kappa_true <- kappa_from_l(l=1700,nu=2)
X <- data.frame(x = 3000*runif(100), y = 3000*runif(100))
dd<- fields::rdist(X,X)
K <- Matern(r=dd,nu=2,var=var_true,kappa=kappa_true)
X$z <-  t(chol(K)) \%*\% rnorm(nrow(X))
var_marg <- var(X["z"])
var_search <- 10^(seq(log10(var_marg/100),log10(var_marg*100),length=100))
rho_search=seq(100,4000,200)
lk_fit <- lscale_from_Matern(X,rho=rho_search,var=var_search,nu=c(2))
print(lk_fit$spat_df)
}

