% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/geometryfns.R
\docType{methods}
\name{manifold}
\alias{manifold}
\alias{manifold,Basis-method}
\title{Manifold}
\usage{
manifold(.Object)

\S4method{manifold}{Basis}(.Object)
}
\description{
Manifold
}

