% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{map_data_to_BAUs}
\alias{map_data_to_BAUs}
\title{Bin data into BAUs}
\usage{
map_data_to_BAUs(data_sp, sp_pols, av_var, variogram.formula = NULL,
  est_error = T)
}
\description{
This is an internal function which bins data into BAUs or aggregates across BAUs if the data have a large footprint. If est_error == T, the observation error is estimated as in Katzfuss & Cressie (2011)
}

