% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Matern}
\alias{Matern}
\title{The Matern function}
\usage{
Matern(r = 0:100, nu = 3/2, var = 1, kappa = 0.1)
}
\arguments{
\item{r}{a vector or matrix of distances.}

\item{nu}{the smoothness parameter.}

\item{var}{the marginal variance}

\item{kappa}{the scale parameters}
}
\value{
a vector or matrix after passing \code{r} through the Matern function.
}
\description{
The Matern function
}
\examples{
x1 <- runif(10)
Sigma <- Matern(as.matrix(dist(x1)),nu=3/2,var=1,kappa=100)
}
\keyword{Matern}
\keyword{covariance}
\keyword{function,}
\keyword{function.}

