% Generated by roxygen2 (4.0.2): do not edit by hand
\name{md5_cache}
\alias{md5_cache}
\title{md5 checksum function}
\usage{
md5_cache(path)
}
\arguments{
\item{path}{the path to where the files will be cached.}
}
\value{
An md5 wrapper for intensive computations. This has arguments
\itemize{
\item \code{fn}: the function being called
\item \code{...}: the arguments to be passed to the function
\item \code{print_output}: if T, details of md5 digest are outputted to screen.
}
}
\description{
Creates a wrapper for generating the md5 checksum for the calling function and its arguments. If a file with the md5 as its filename exists in the specified folder then this is loaded. Otherwise the function is evaluated and the results are stored in the specified folder.
}
\details{
md5 checksums on the called function and any function-arguments are generated by creating a text file in the specified folder and digesting
the text file before deleting it. md5 checksums on character arguments are carried out on the file (if the file exists) or on the character as appropriate.
}
\examples{
myfun1 <- function(x) x^2
myfun2 <- function(y,f) y^3 + f(y)
md5_wrapper <- md5_cache(".")
x <- md5_wrapper(myfun2,2,myfun1)
}
\keyword{digest}
\keyword{md5,}

