% Generated by roxygen2 (4.0.2): do not edit by hand
\name{md5_fun}
\alias{md5_fun}
\title{Generate md5 checksum from function}
\usage{
md5_fun(fn, path)
}
\arguments{
\item{fn}{the function to digest}

\item{path}{the directory where dumping will be carried out}
}
\description{
Creates an md5 checksum of a function by dumping it and then digesting the file.
}
\details{
Note that a ~/cache folder needs to be present for this to work.
}
\examples{
myfun1 <- function(x) x^2
x <- md5_fun(myfun1,".")
}
\keyword{digest}
\keyword{md5,}

