% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{measure-class}
\alias{measure-class}
\title{measure}
\description{
An object of class measure contains a distance function and a variable \code{dim} with the dimensions of the Riemannian manifold over which the distance is computed. By default, the distance function used is inherited from the package fields.
}
\details{

}
\keyword{Manifolds,}
\keyword{planes}
\keyword{spheres,}

