% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{MODIS_cloud_df}
\alias{MODIS_cloud_df}
\title{MODIS cloud data}
\format{
A data frame with 33,750 rows and 3 variables:
\describe{
  \item{x}{x-coordinate}
  \item{y}{y-coordinate}
  \item{z}{binary dependent variable: 1 if cloud is present, 0 if no cloud. This
  variable has been thresholded from the original continuous measurement of
  radiance supplied by the MODIS instrument}
  \item{z_unthresholded}{The original continuous measurement of
  radiance supplied by the MODIS instrument}
}
}
\usage{
MODIS_cloud_df
}
\description{
An image of a cloud taken by the Moderate Resolution
Imaging Spectroradiometer (MODIS) instrument aboard the Aqua satellite (MODIS
Characterization Support Team, 2015).
}
\references{
MODIS Characterization Support Team (2015). MODIS 500m Calibrated Radiance Product.NASA MODIS Adaptive Processing System, Goddard Space Flight Center, USA.
}
\keyword{datasets}
