% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/basisfns.R
\docType{methods}
\name{nbasis}
\alias{nbasis}
\alias{nbasis,Basis-method}
\alias{nbasis,SRE-method}
\title{Number of basis functions}
\usage{
nbasis(.Object)

\S4method{nbasis}{Basis}(.Object)

\S4method{nbasis}{SRE}(.Object)
}
\description{
Number of basis functions
}

