% Generated by roxygen2 (4.0.2): do not edit by hand
\name{neighb_from_prec}
\alias{neighb_from_prec}
\title{Return neighb list from precision structure}
\usage{
neighb_from_prec(Q)
}
\arguments{
\item{Q}{a sparse matrix}
}
\value{
a list of vectors as described in the description
}
\description{
Takes a precision matrix and returns a list of vectors, where the \code{i}th vector contains \code{i} and the indices of the neighbours of the \code{i}th variable.
}
\examples{
G <- GMRF_RW()
nlist <- neighb_from_prec(getPrecision(G))
}

