% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{new_opts_FRK}
\alias{new_opts_FRK}
\title{Options for FRK package}
\usage{
new_opts_FRK(d = list(progress = TRUE, verbose = FALSE, parallel = 6L, Rhipe =
  FALSE))
}
\description{
Options for FRK package
}

