% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{NOAA_df_1990}
\alias{NOAA_df_1990}
\title{NOAA maximum temperature data for 1990-1993}
\format{A data frame with 196253 rows and 8 variables:
\describe{
  \item{year}{year of retrieval}
  \item{month}{month of retrieval}
  \item{day}{day of retrieval}
  \item{z}{dependent variable}
  \item{proc}{variable name (Tmax)}
  \item{id}{station id}
  \item{lat}{latitude coordinate of retrieval}
  \item{lon}{longitude coordinate of retrieval}
}}
\usage{
NOAA_df_1990
}
\description{
Maximum temperature data obtained from the National Oceanic and Atmospheric Administration (NOAA) for a part of the USA between 1990 and 1993 (inclusive).
}
\references{
Chahine, M. et al. (2006). AIRS: Improving weather forecasting and providing new data on greenhouse gases. Bulletin of the American Meteorological Society 87, 911--26.
}
\keyword{datasets}

