% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nrow}
\alias{nrow}
\alias{nrow,FEBasis-method}
\alias{nrow,GMRF-method}
\alias{nrow,GMRF_basis-method}
\alias{nrow,Obs-method}
\title{Return size of object Get stiffness matrix}
\usage{
nrow(x)

\S4method{nrow}{GMRF}(x)

\S4method{nrow}{GMRF_basis}(x)

\S4method{nrow}{FEBasis}(x)

\S4method{nrow}{Obs}(x)
}
\arguments{
\item{x}{an MVST object}
}
\description{
Returns the size of the object, where the 'size' is determined by the class of the object being passed. For a \code{GMRF} and \code{GMRF_basis}, this is the number of variables of the associated MRF, for an \code{FEBasis} this is the number of vertices in the mesh, and for an \code{Obs} this is the number of observations.
}
\examples{
data(icesat)
icesat_obs <- Obs(df=icesat)
nrow(icesat_obs)
}

