% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Obs}
\alias{Obs}
\title{Observation block}
\usage{
Obs(df, name = "Obs", remove_cross_ins = 0, ...)
}
\arguments{
\item{df}{a data frame which should contain at least 5 entries, \code{x,y,t,z} and \code{std} which denote the horizontal, vertical and temporal indices of the observations, the value and error respectively. Alternatively this could be a path name.}

\item{name}{the name of the observation process}

\item{remove_cross_ins}{removes data outside a circle centred at zero with specified radius. Convenient when working with satellite data in polar stereographic projection when some cross-ins are detected.}

\item{...}{other arguments passed on to \code{preprocess_obs}}
}
\value{
Object of class \code{Obs} (which inherits from class \code{block} and is thus also a block)
}
\description{
This function initialises an object of class \code{Obs} which defines a an observation data set. By default, this is for observations with negligible spatial footprint. For larger supports, use \code{Obs_poly}.
}
\examples{
O <- Obs(df=data.frame(x=runif(5),y=runif(5),t=c(1,1,1,2,2),z=runif(5),std=runif(5)))
print(O)
plot(subset(O,t==1),"z",pt_size=4)
}
\keyword{Observations,}
\keyword{block}
\keyword{change}
\keyword{of}
\keyword{support,}

