% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Obs_poly}
\alias{Obs_poly}
\title{Observation block with support}
\usage{
Obs_poly(pol_df, name = "Obs_poly", alpha0 = NA, av_dist = NA, ...)
}
\arguments{
\item{pol_df}{a wide table format data frame identifying support of observation, or the path to a file containing the data. The data should be in wide-table format and hava column denoting the observation \code{id}, vertices \code{x1, y1, x2, y2, ...} and the time point \code{t} if required.}

\item{alpha0}{sets, if needed, an averaging matrix over observations. If this specified, also a parameter \code{av_dist} needs to be specified}

\item{av_dist}{denotes within which distance observations are considered neighbours.}

\item{...}{further arguments passed to the parent \code{Obs()} constructor for further processing.}
}
\value{
Object of class \code{Obs_poly} (which inherits from class \code{Obs} and is thus also an observation block)
}
\description{
This function initialises an object of class \code{Obs_poly} which defines a an observation data setand associated spatial supports. IMPORTANT: The vertices need to be in consecutive order.
}
\examples{
# Create a polygon 'footprint'
pol_df <- data.frame(id=1,x1=0,x2=0,x3=1,x4=1,y1=0,y2=1,y3=1,y4=0,t=0)
pol_df <- rbind(pol_df,data.frame(id=2,x1=-0.5,x2=-0.5,x3=0.5,x4=0.5,y1=-0.5,y2=0.5,y3=0.5,y4=-0.5,t=0))
df <- data.frame(id=1,x=0.5,y=0.5,z=1,std=1,t=0)
df <- rbind(df,data.frame(id=2,x=0,y=0,z=0.2,std=0.2,t=0))
O <- Obs_poly(df=df,pol_df=pol_df)
plot(O,"z")
}
\keyword{Observations,}
\keyword{block}
\keyword{change}
\keyword{of}
\keyword{support,}

