% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Obs_poly-class}
\alias{Obs_poly-class}
\title{Observations with non-trivial spatial footprints}
\description{
This class inherits from \code{Obs}. It extends \code{Obs} by providing the ability to encode spatial footprints with an additional input.
The use of this class requires installation of \code{gpclib}.
}
\details{

}

