% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{observed_BAUs}
\alias{observed_BAUs}
\alias{observed_BAUs,SRE-method}
\title{Indices of observed BAUs}
\usage{
observed_BAUs(SRE_model)

\S4method{observed_BAUs}{SRE}(SRE_model)
}
\arguments{
\item{SRE_model}{object of class \code{SRE}}
}
\value{
The indices (a numeric vector) of the observed BAUs
}
\description{
Takes an object of class \code{SRE} and returns the indices of the observed BAUs
}
\seealso{
See \code{\link{unobserved_BAUs}}.
}
