% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partition_nei_from_points}
\alias{partition_nei_from_points}
\title{Find cpartition neighbourhood from underlying graph}
\usage{
partition_nei_from_points(points, nei_list)
}
\arguments{
\item{points}{a data frame with fields \code{x}, \code{y} and \code{class}}

\item{nei_list}{a list where each element is a vector containing the indices of the neighbours of the corresponding node}
}
\value{
a neighbourhood list of the superset
}
\description{
This function takes a set of points with a field \code{class} and the neighbourhood list of those points,
and computes the neighbourhood list of the super-graph composed by treating the \code{class} as a vertex in its own right. If even
one underlying vertex has a nerghbour in another partition, then that partition is the present partition's neighbour.
}
\author{
Andrew Zammit Mangion
}

