% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partition_section}
\alias{partition_section}
\title{Recursion for the Kernigham-Lin equations}
\usage{
partition_section(points, nrep)
}
\arguments{
\item{points}{a data frame with fields \code{x}, \code{y} and \code{id}, where \code{id} typically ranges from 1 to the number of points in the graph}

\item{nrep}{the number of times to split the graph}
}
\value{
a list of vectors, where each vector contains the \code{id}s of the nodes in a particular group
}
\description{
The  Kernigham-Lin equations split a graph into two equal parts depending on a weighting matrix. The weighting
matrix here is constrained to be the Eucledian distance between the points.
}
\examples{
require(ggplot2)
N = 99
points <- data.frame(x = runif(n=N), y = runif(n=N),id=1:N)
points <- partition_section(points,nrep=4)
ggplot() + geom_point(data=points,aes(x,y,colour=class,size=4))
}
\author{
C. Ladroue, A. Zammit Mangion
}

