% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{plane}
\alias{plane}
\title{plane}
\usage{
plane(measure = Euclid_dist(dim = 2L))
}
\arguments{
\item{measure}{an object of class \code{measure}}
}
\description{
Initialisation of a 2D plane.
}
\details{
A 2D plane is initialised using a \code{measure} object. By default, the measure object (\code{measure}) is the Euclidean distance in 2 dimensions, \link{Euclid_dist}.
}
\examples{
P <- plane()
print(type(P))
print(sp::dimensions(P))
}
