% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{plot,SRE,ANY-method}
\alias{plot,SRE,ANY-method}
\title{Plot predictions from FRK analysis.}
\usage{
\S4method{plot}{SRE,ANY}(x, y, zdf = NULL, map_layer = NULL, subset_time = NULL, ...)
}
\arguments{
\item{x}{\code{SRE} object}

\item{y}{result of calling \code{predict} on an \code{SRE} object}

\item{zdf}{a \code{data.frame}, \code{SpatialPointsDataFrame}, or \code{SpatialPolygonsDataFrame} containing the observations}

\item{map_layer}{(optional) a \code{ggplot} layer or object to add below the plotted layer, often a map}

\item{subset_time}{(optional) a vector of times to be included; applicable only for \code{STFDF} objects}

\item{...}{optional arguments passed on to whatever geom is appropriate for the data (geom_point, geom_raster, or geom_polygon)}
}
\value{
a list of \code{ggplot} objects consisting of the observed data, predictions, and standard errors. This list can then be supplied to, for example, \code{ggpubr::ggarrange()}.
}
\description{
Plot predictions from FRK analysis.
}
\seealso{
\code{\link{plot_spatial_or_ST}}
}
