% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/plottingfns.R
\name{plot}
\alias{plot}
\alias{plot,SRE,ST-method}
\alias{plot,SRE,Spatial-method}
\title{Plot predictions from FRK analysis}
\usage{
plot(x, y, ...)

\S4method{plot}{SRE,ST}(x, y, ...)

\S4method{plot}{SRE,Spatial}(x, y, ...)
}
\arguments{
\item{x}{object of class \code{SRE}}

\item{y}{the \code{Spatial*DataFrame} or \code{STFDF} object resulting from the call \code{predict(x)}}

\item{...}{optional arguments passed on to \code{\link{plot_spatial_or_ST}}}
}
\value{
A list of \code{ggplot} objects consisting of the observed data, predictions, and standard errors. This list can then be supplied to, for example, \code{ggpubr::ggarrange()}.
}
\description{
This function acts as a wrapper around 
\code{\link{plot_spatial_or_ST}}, using the \code{SRE} object \code{x} to 
determine which fields of \code{y} should be plotted, and the \code{@data} 
slot of \code{x} to plot the supplied data. It also generates 
some informative, latex-style legend labels for each of the plots.
}
\examples{
## See example in the help file for SRE
}
