% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{plot_spatial_or_ST}
\alias{plot_spatial_or_ST}
\title{Plot data from a Spatial*DataFrame or STFDF object}
\usage{
plot_spatial_or_ST(
  object,
  column_names,
  map_layer = NULL,
  subset_time = NULL,
  uncertainty = FALSE,
  ...
)
}
\arguments{
\item{object}{the Spatial*DataFrame or STFDF we wish to plot}

\item{column_names}{a vector of strings indicating the columns of the data to plot}

\item{map_layer}{(optional) a \code{ggplot} layer or object to add below the plotted layer, often a map}

\item{subset_time}{(optional) a vector of times to be included; applicable only for \code{STFDF} objects}

\item{uncertainty}{a logical indicating whether an uncertainty (green and brown) colour/fill scale is to be used; default is FALSE, so that a Spectral palette is used}

\item{...}{optional arguments passed on to whatever geom is appropriate for the data (geom_point, geom_raster, or geom_polygon)}
}
\description{
Plot data from a Spatial*DataFrame or STFDF object
}
