% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{plotting-themes}
\alias{EmptyTheme}
\alias{LinePlotTheme}
\alias{plotting-themes}
\title{Plotting themes}
\usage{
LinePlotTheme()

EmptyTheme()
}
\value{
Object of class \code{ggplot}
}
\description{
Formats a ggplot object for neat plotting.
}
\details{
\code{LinePlotTheme()} creates \code{ggplot} object with a white background, a relatively large font and grid lines. \code{EmptyTheme()} on the other hand creates a \code{ggplot} object with no axes or legends.
}
\examples{
\dontrun{
X <- data.frame(x=runif(100),y = runif(100), z = runif(100))
LinePlotTheme() + geom_point(data=X,aes(x,y,colour=z))
EmptyTheme() + geom_point(data=X,aes(x,y,colour=z))}
}
\keyword{ggplot}

