% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Poisson_simulated}
\alias{Poisson_simulated}
\title{Simulated spatial count (Poisson) data}
\format{
A data frame with 750 rows and 3 variables:
\describe{
  \item{Z}{Response}
  \item{x}{Spatial x-ordinate}
  \item{y}{Spatial y-ordinate}
}
}
\usage{
Poisson_simulated
}
\description{
Simulated spatial count (Poisson) data
}
\keyword{datasets}
