% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Prec_from_SPDE_wrapper}
\alias{Prec_from_SPDE_wrapper}
\title{Precision matrix from SPDE}
\usage{
Prec_from_SPDE_wrapper(M, K, nu, desired_prec, l)
}
\arguments{
\item{M}{the mass matrix of the finite element basis}

\item{K}{the stiffness matrix of the finite element basis}

\item{nu}{the smoothness parameter of the Matern kernel}

\item{desired_prec}{a constant or a vector of length \code{n} where each element contains the
precision (1 over the marginal standard deviation squared) at each vertex.}

\item{l}{a constant or a vector of length \code{n} where each element contains the
practical range (distance at which correlation is 0.1) at each vertex. This should be spatially smooth. Informally,
this should vary slower than the field itself, otherwise a good local approximation to the Matern representation is not guaranteed.}
}
\value{
Object of class "dgCMatrix"
}
\description{
This function returns the precision matrix for a GMRF over a triangulation such that it is an approximation
to a Gaussian field with a Matern kernel with pre-specified parameters.
}
\examples{
require(Matrix)
data(surf_fe)
mu <- matrix(0,nrow(surf_fe$p),1)
Q <- Prec_from_SPDE_wrapper(surf_fe$M,
                            surf_fe$K,
                            nu = 2,
                            desired_prec = 1,
                            l = 100)
my_GMRF <- GMRF(mu=mu, Q=Q,name="my_first_GMRF")
}
\keyword{GMRF}

