% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{pred_variance_large}
\alias{pred_variance_large}
\alias{pred_variance_large,list,Graph_2nodes-method}
\alias{pred_variance_large,list-Graph_2nodes-method}
\alias{pred_variance_large,matrix,Graph_2nodes-method}
\alias{pred_variance_large,matrix-Graph_2nodes-method}
\title{Predictive variance for large data sets}
\usage{
pred_variance_large(Results, G)

\S4method{pred_variance_large}{list,Graph_2nodes}(Results, G)

\S4method{pred_variance_large}{matrix,Graph_2nodes}(Results, G)
}
\arguments{
\item{Results}{a list generated by the function \code{Infer}}

\item{G}{An object of class \code{Graph_2nodes} describing the relationship between the observations and the processes.}
}
\value{
Object of class \code{Obs}.
}
\description{
This function takes inferential results from \code{Infer} and computes the predictive variance from the posterior GMRF at locations for a very large number of data points. This function only yields valid results if the non-zero locations in each row of the incidence matrix constructed using the validation data are a subset of those used to generate the posterior precision matrix, see references.
}
\examples{
\dontrun{
require(Matrix)
data(icesat)
data(surf_fe)

## First create observation object
icesat_obs <- Obs(df=icesat,
                 abs_lim = 5,
                 avr_method = "median",
                 box_size=100,
                 name="icesat")

## Now create GMRF defined over some FE basis
Mesh <- initFEbasis(p=surf_fe$p,
                    t=surf_fe$t,
                    M=surf_fe$M,
                    K=surf_fe$K)

mu <- matrix(0,nrow(Mesh),1)
Q <- sparseMatrix(i=1:nrow(surf_fe$p), j = 1:nrow(surf_fe$p), x = 1)

my_GMRF <- GMRF(mu = mu, Q = Q,name="SURF",t_axis = 0:6)
SURF <-GMRF_basis(G = my_GMRF, Basis = Mesh)

L1 <- link(SURF,icesat_obs)
e <- link_list(list(L1))
v <- block_list(list(O = icesat_obs, G = SURF))
G <- new("Graph",e=e,v=v)
G_reduced <- compress(G)
Results <- Infer(G_reduced)

Obs_test <- pred_variance_large(Results,G_reduced)
}
}
\references{
Jonathan C. Rougier, Andrew Zammit-Mangion and Nana Schoen (2014). Computation and visualisation for large-scale Gaussian updates. \url{http://arxiv.org/abs/1406.5005}.
}
\keyword{predictive}
\keyword{validation}
\keyword{variance,}

