% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess_obs}
\alias{preprocess_obs}
\title{Pre-processes observations}
\usage{
preprocess_obs(Obs.obj, std_lim = NA, abs_lim = NA, avr_method = NA,
  box_size = 10, min_pts = 4, ...)
}
\arguments{
\item{Obs.obj}{The observation object}

\item{std_lim}{Values with error higher than \code{std_lim} are deleted}

\item{abs_lim}{Values which are bigger than \code{abs_lim} are deleted}

\item{avr_method}{If \code{mean} then the mean value and mean error on a sub-grid are used to sub-sample the observations. If \code{median} the median value and MAD of the z-values are used
for subsampling. If \code{mean_no_std} or \code{median_no_std}, the error is not computed ignored.}

\item{box_size}{The grid width over which the observations are averaged}

\item{min_pts}{If a grid box contains less than \code{min_pts} it is marked as empty}

\item{...}{Other parameters which are ignored}
}
\description{
This function simply takes a data frame and performs standard pre-processing functions such as removing obvious outliers and averaging data
in computationally efficient way over a regular grid
}

