% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RBF_filter}
\alias{RBF_filter}
\title{RBF filter}
\usage{
RBF_filter(data, si, varname = "z", smooth_var = 800^2)
}
\arguments{
\item{data}{a data frame with fields \code{x}, \code{y} and an output variable.}

\item{si}{data frame containing fields \code{x} and \code{y} which denote the desired coordinates of the output field (typically a grid).}

\item{varname}{the label of the column in \code{data} to use as the output variable.}

\item{smooth_var}{the variance of the radial basis function.}
}
\value{
the smoothed field  values at locations in \code{si}
}
\description{
Takes a 2-D set of disaggregated points and smooths them out by placing a radial basis function at each point and using
least-squares estimation for the ensuing weights. The constant trend is added on at the end.
}
\examples{
data <- data.frame(x = c(1,1,2,2),y = c(1,2,1,2), z = c(1,2,3,4))
si <- as.data.frame(expand.grid(seq(0,3,by=0.1),seq(0,3,by=0.1)))
names(si) <- c("x","y")
si$z <- RBF_filter(data,si,varname="z",smooth_var=1)
}

