% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{real_line}
\alias{real_line}
\title{real line}
\usage{
real_line(metric = Euclid_dist(dim = 1L))
}
\arguments{
\item{metric}{an object of class \code{measure}}
}
\description{
Initialisation of the 1D real-line manifold.
}
\details{
A real line is initialised using a \code{measure} object. By default, the measure object (\code{metric}) describes the distance between two points as the absolute difference between the two coordinates.
}
\examples{
R <- real_line()
print(type(R))
print(sp::dimensions(R))
}

