% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{remove_BAUs}
\alias{remove_BAUs}
\alias{remove_BAUs,SpatialPixelsDataFrame-method}
\alias{remove_BAUs,STFDF-method}
\title{Remove BAUs}
\usage{
remove_BAUs(BAUs, rmidx)

\S4method{remove_BAUs}{SpatialPixelsDataFrame}(BAUs, rmidx)

\S4method{remove_BAUs}{STFDF}(BAUs, rmidx)
}
\arguments{
\item{BAUs}{BAU object.}

\item{rmidx}{indices of spatial basis functions to remove. For spatio-temporal BAUs, the indices of the spatial BAUs (\code{BAUs@sp}) must be provided.}
}
\value{
\code{BAUs} with specified BAUs removed.
}
\description{
Takes a \code{BAU} object returns the \code{BAU} object with selected BAUs removed specified by \code{rmidx}.
For spatio-temporal BAUs, the indices of the spatial BAUs (\code{BAUs@sp}) must be provided, and the spatio-temporal BAUs corresponding to these spatial BAUs will be removed.
}
