% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{reverse_BAU_coords}
\alias{reverse_BAU_coords}
\alias{reverse_BAU_coords,SpatialPixelsDataFrame-method}
\title{Reverses BAU coordinates}
\usage{
reverse_BAU_coords(BAUs)

\S4method{reverse_BAU_coords}{SpatialPixelsDataFrame}(BAUs)
}
\arguments{
\item{BAUs}{An object of class \code{SpatialPixelsDataFrame}.}
}
\value{
\code{BAUs} with the reversed coordinate order.
}
\description{
Reverses the coordinate order (and all relevant slots) of a spatial BAU object (an object of class \code{SpatialPixelsDataFrame}). This is useful when comparing BAU objects to \code{SpatialPolygon} objects, for instance, when one wishes to trim BAUs based on some polygon.
}
\examples{
library(sp)
RNGversion("3.6.0")
set.seed(1)
data <- data.frame(lon = runif(10, 0, 10), 
                   lat = runif(10, 0, 40), 
                   z= rnorm(10))
coordinates(data) <- ~ lon + lat
BAUs <- auto_BAUs(manifold = plane(), 
                  cellsize = 0.5, 
                  data=data)
\dontrun{plot(BAUs)}
BAUs <- reverse_BAU_coords(BAUs)
\dontrun{plot(BAUs)}
}
