% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{reverse_spatial_coords}
\alias{reverse_spatial_coords}
\title{Reverses spatial coordinates}
\usage{
reverse_spatial_coords(BAUs)
}
\arguments{
\item{BAUs}{An object of class \code{SpatialPixelsDataFrame}, \code{SpatialPolygonsDataFrame}, or \code{STFDF}.}
}
\value{
\code{BAUs} with the reversed coordinate order.
}
\description{
Reverses the coordinate order (and all relevant slots) of a BAU object. This is useful when comparing BAU objects to \code{SpatialPolygon} objects, for instance, when one wishes to trim BAUs based on some polygon.
}
\examples{
library(sp)
RNGversion("3.6.0")
set.seed(1)
data <- data.frame(lon = runif(10, 0, 10), 
                   lat = runif(10, 0, 40), 
                   z   = rnorm(10))
## Create SpatialPixelsDataFrame:
coordinates(data) <- ~ lon + lat
BAUs <- auto_BAUs(manifold = plane(), 
                  cellsize = 0.5, 
                  data=data)
BAUs@bbox # lon comes before lat
\dontrun{plot(BAUs)}
BAUs <- reverse_spatial_coords(BAUs)
BAUs@bbox # lat comes before lon
\dontrun{plot(BAUs)}
}
