% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roughness_from_topo}
\alias{roughness_from_topo}
\title{Extract roughness from topography}
\usage{
roughness_from_topo(df, ds = 20)
}
\arguments{
\item{df}{data frame with columns \code{x, y} and \code{z}}

\item{ds}{the grid size required}
}
\description{
Supplied with a data frame with fields x, y and z, this function bins the data into boxes of length \code{ds} and returns a data frame where \code{z} is now the estimated roughness, given by the lof of the standard deviation of the topography in the boxes.
}
\examples{
df <- data.frame(x=runif(1000),y=runif(1000),z=runif(1000))
df_roughness <- roughness_from_topo(df,ds=0.1)
}

