% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_GMRF}
\alias{sample_GMRF}
\title{Sample from a GMRF}
\usage{
sample_GMRF(G, L = NULL, reps = 1, P = NULL)
}
\arguments{
\item{G}{object of class \code{GMRF}}

\item{L}{Cholesky factor of the precision matrix, if \code{P} is \code{NULL} then this is treated as the factor of an unpermuted matrix}

\item{reps}{number of samples to generate}

\item{P}{permutation matrix}
}
\value{
a matrix of size \code{n} by \code{reps}
}
\description{
Takes a GMRF object and, possibly, associated permutation matrix and Cholesky factor of permuted precision matrix, to generate
samples.
}
\examples{
G <- GMRF_RW()
G <- setPrecision(G, getPrecision(G) + 0.01*Imat(nrow(G)))
X <- sample_GMRF(G,reps=10)
}

