% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{shapefiles}
\alias{shapefiles}
\title{Boundaries for Antarctica}
\format{A list with three fields. Each data frame has at least the following fields
 \itemize{
  \item x. x coordinate in km under a polar stereographic projection
  \item y. y coordinate in km under a polar stereographic projection
  \item id. polygon id
}}
\description{
A list of data frames containing useful shapefiles for plotting results relating to Antarctica.
}
\details{
\itemize{
  \item grounding_sub. a 1984 x 2 matrix containing the vertices
  \item coast_sub. a 3911 x 3 matrix containing the trianulations
  \item Islands. a 1984 x 1984 sparse mass matrix
  \item K. a 1984 x 1984 sparse stiffness matrix
}
}
\keyword{datasets}

