% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/plottingfns.R
\docType{methods}
\name{show_basis}
\alias{show_basis}
\alias{show_basis,ANY,Basis-method}
\alias{show_basis,Basis-method}
\title{Show basis functions}
\usage{
show_basis(g, basis, ...)

\S4method{show_basis}{ANY,Basis}(g = ggplot(), basis)
}
\arguments{
\item{g}{object of class \code{gg} (a \code{ggplot} object)}

\item{basis}{object of class \code{Basis}}
}
\description{
Generic plotting function for displaying the basis functions
}
\details{
The function \code{show_basis} adapts its behaviour to the manifold being used. With \code{real_line}, the one-dimensional basis functions are plotted with colour distinguishing between the different resolutions. With \code{plane}, only radial basis functions are supported (at present). Each basis function is shown as a circle with diameter is equal to the \code{scale} parameter of the function. Linetype distinguishes the resolution. With \code{sphere}, the the centres of the basis functions are show as circles, with larger sizes corresponding to lower resolutions.
}
\examples{
library(ggplot2)
library(sp)
data(meuse)
coordinates(meuse) = ~x+y # change into an sp object
G <- auto_basis(m = plane(),data=meuse,nres = 2,regular=2,prune=10,type = "Gaussian")
show_basis(ggplot(),G) + geom_point(data=data.frame(meuse),aes(x,y))
}

