% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{sphere}
\alias{sphere}
\title{sphere}
\usage{
sphere(radius = 6371)
}
\arguments{
\item{radius}{radius of sphere}
}
\description{
Initialisation of the 2-sphere, S2.
}
\details{
The 2D surface of a sphere is initialised using a \code{radius} parameter. The default value of the radius \code{R} is \code{R}=6371 km, Earth's radius, while the measure used to compute distances on the sphere is the great-circle distance on a sphere of radius \code{R}.
}
\examples{
S <- sphere()
print(sp::dimensions(S))
}
