% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{split_validation}
\alias{split_validation}
\alias{split_validation,Obs-method}
\title{Split data into training/testing groups}
\usage{
split_validation(.Object, samples, common, ...)

\S4method{split_validation}{Obs}(.Object, samples, common = 0, ...)
}
\arguments{
\item{.Object}{an object of class \code{Obs}.}

\item{samples}{an integer identifying the number of observations to use for validation.}

\item{common}{a flag which, if 1, indicates that validation data should be chosen to coincide spatially. This flag is not relevant if the
data is not spatio-temporal.}

\item{...}{further arguments passed on to \code{subset}.}
}
\value{
a list of two objects of class \code{Obs}.
}
\description{
\code{split_validation} takes an object of class \code{Obs} and returns a list of two objects of class \code{Obs}, where one object
contains the training data and the other contains the validation data. If the data is spatio-temporal, the validation data can be chosen to be
co-located in space.
}
\examples{
data(icesat)
icesat_obs <- Obs(df=icesat)
O2 <- split_validation(icesat_obs,100,common=0, t > 0)
}

