% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SRE-class}
\alias{SRE-class}
\title{SRE model}
\description{
SRE model
}
\details{

}
\keyword{Spatial}
\keyword{effects,}
\keyword{fixed}
\keyword{kriging}
\keyword{random}
\keyword{rank}

