% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SREutils.R
\name{SRE.fit}
\alias{SRE.fit}
\title{Estimate SRE model parameters}
\usage{
SRE.fit(SRE_model, n_EM = 100L, tol = 1e-05, method = "EM",
  print_lik = F)
}
\description{
Estimate SRE model parameters
}

