% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{SRE.plot}
\alias{SRE.plot}
\title{Plot predictions from FRK analysis}
\usage{
SRE.plot(SRE_model, Pred, zdf = NULL)
}
\arguments{
\item{SRE_model}{\code{SRE} object}

\item{Pred}{result of a calling \code{predict} on \code{SRE_model}}

\item{zdf}{a \code{data.frame} or \code{SpatialPointsDataFrame} containing the observations}
}
\value{
A list of \code{ggplot} objects consisting of the observed data, predictions, and standard errors. This list can then be supplied to, for example, \code{ggpubr::ggarrange()}.
}
\description{
Plot predictions from FRK analysis
}
