% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SREpredict.R
\name{SRE.predict}
\alias{SRE.predict}
\title{Deprecated: Please use \code{\link{predict}}}
\usage{
SRE.predict(...)
}
\arguments{
\item{...}{(Deprecated)}
}
\description{
Deprecated: Please use \code{\link{predict}}
}
