% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{STplane}
\alias{STplane}
\title{plane in space-time}
\usage{
STplane(metric = Euclid_dist(dim = 3L))
}
\arguments{
\item{metric}{an object of class \code{measure}}
}
\description{
Space-time plane initialisation.
}
\details{
A 2D plane with a time component added is initialised using a \code{measure} object. By default, the measure object (\code{metric}) is the Euclidean distance in 3 dimensions, \link{Euclid_dist}. Distances are computed using functions from the \code{fields} package.
}
\examples{
P <- STplane()
print(type(P))
warning("CHECK: This may not work for ST plane")
print(sp::dimensions(P))
}

