% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{subset}
\alias{subset}
\alias{subset,Basis-method}
\alias{subset,GMRF-method}
\alias{subset,GMRF_basis-method}
\alias{subset,Obs-method}
\alias{subset,Obs_poly-method}
\title{Subset an MVST object}
\usage{
subset(x, ...)

\S4method{subset}{Basis}(x, ...)

\S4method{subset}{GMRF_basis}(x, ...)

\S4method{subset}{GMRF}(x, ...)

\S4method{subset}{Obs}(x, ...)

\S4method{subset}{Obs_poly}(x, ...)
}
\arguments{
\item{x}{the MVST object}

\item{...}{relations pertaining to the data frame within the MVST object}
}
\value{
object of the same class as \code{x} unless \code{x} is a mesh, in which case a data frame is returned.
}
\description{
This function can be used to subset observations, meshes, and latent fields in the usual manner of \code{subset}.
}
\examples{
data(icesat)
icesat_obs <- Obs(df=icesat)
icesat_sub <- subset(icesat_obs,t==3)
print(class(icesat_sub))

data(surf_fe)
Mesh <- initFEbasis(p=surf_fe$p, t = surf_fe$t, M = surf_fe$M, K = surf_fe$K)
Mesh_sub <- subset(Mesh,x < 0)
print(class(Mesh_sub))
}

