% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{surf_fe}
\alias{surf_fe}
\title{Example mesh}
\format{A list with four fields}
\description{
A list containing the mesh details of the Antarctic ice sheet at moderate (~200km) resolution
}
\details{
\itemize{
  \item p. a 1984 x 2 matrix containing the vertices
  \item t. a 3911 x 3 matrix containing the trianulations
  \item M. a 1984 x 1984 sparse mass matrix
  \item K. a 1984 x 1984 sparse stiffness matrix
}
}
\keyword{datasets}

