% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/basisfns.R
\docType{methods}
\name{TensorP}
\alias{TensorP}
\alias{TensorP,Basis,Basis-method}
\alias{TensorP,Basis-Basis-method}
\title{Tensor product of basis functions}
\usage{
TensorP(Basis1, Basis2)

\S4method{TensorP}{Basis,Basis}(Basis1, Basis2)
}
\arguments{
\item{Basis1}{first set of basis functions}

\item{Basis2}{second set of basis functions}
}
\description{
Constructs a new set of basis by finding the tensor product of two sets of basis functions. The product is carried out on the basis functions elementwise. Hence, \code{Basis1} and \code{Basis2} need to contain the same number of basis functions.
}

