% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{timeline}
\alias{timeline}
\title{timeline}
\usage{
timeline(measure = Euclid_dist(dim = 1L))
}
\arguments{
\item{measure}{an object of class \code{measure}}
}
\description{
Initialisation of a timeline (real line).
}
\details{
A time axis initialised using a \code{measure} object. By default, the measure object (\code{measure}) is the absolute difference.
}
\examples{
P <- timeline()
print(type(P))
print(sp::dimensions(P))
}

