% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{unobserved_BAUs}
\alias{unobserved_BAUs}
\alias{unobserved_BAUs,SRE-method}
\title{Indices of unobserved BAUs}
\usage{
unobserved_BAUs(SRE_model)

\S4method{unobserved_BAUs}{SRE}(SRE_model)
}
\arguments{
\item{SRE_model}{object of class \code{SRE}}
}
\value{
The indices (a numeric vector) of the unobserved BAUs
}
\description{
Takes an object of class \code{SRE} and returns the indices of the unobserved BAUs
}
\seealso{
See \code{\link{observed_BAUs}}.
}
