% Generated by roxygen2 (4.0.2): do not edit by hand
\name{VAR_Gauss}
\alias{VAR_Gauss}
\title{Vector auto-regressive model with GMRF represenation}
\usage{
VAR_Gauss(mu_fun = function(k) return(matrix(0, 2, 5)), A_fun = function(k)
  return(Imat(2)), B_fun = function(k) return(emptySp()),
  Qw_fun = function(k) return(Imat(2)), t_axis = c(0:6), Qb = emptySp(),
  name = "none")
}
\arguments{
\item{mu_fun}{function of time \eqn{k}, returns matrix of size \eqn{n}}

\item{A_fun}{function of time \eqn{k}, returns sparse matrix of size \eqn{n\times n}}

\item{B_fun}{function of time \eqn{k}, returns  sparse matrix of size \eqn{n\times m}}

\item{Qw_fun}{function of time \eqn{k}, returns  sparse matrix of size \eqn{n\times n}}

\item{t_axis}{time axis of process}

\item{Qb}{prior precision matrix of \eqn{\beta}; sparse matrix of size \eqn{m \times m}}

\item{name}{name of VAR}
}
\value{
Object of class VAR_Gauss which inherits from class \code{GMRF}.
}
\description{
This function initialises a vector auto-regressive model and represents it as a Gaussian Markov Random Field with mean \code{mu} and precision matrix \code{Q}. This constructor differs from other GMRF constructors in that it takes function inputs
to define temporally evolving characteristics. The default representation is \deqn{x_{k+1} = \mu_k + A_kx_k + B_k\beta_k + e_k} where
\eqn{e_k \sim \mathcal{N}(0,Q_k)}. Note that in addition to covariates, a known mean \eqn{\mu_k} can be added, this can be omitted and
replaced appropriately with entries in \eqn{B_k}. A multi-variate vector auto-regressive model can be speficied by letting
\code{A_fun} and \code{Qw_fun} return matrices a multiple of the dimension of the underlying basis over which the GMRF is defined.
}
\examples{
require(Matrix)
t_axis <- 0:10
mu <- function(k) return(matrix(0,length(t_axis),1))
A <- function(k)  return(sparsediag(0.4))
B <- function(k)  cBind(Imat(1),k*Imat(1))
Q <- function(k)  return(sparsediag(1))
Qb = bdiag(Imat(1),Imat(1))
VAR <- VAR_Gauss( mu_fun = mu,A=A, B=B, Qw = Q,t_axis = t_axis,Qb=Qb,name="firstVAR")
}
\keyword{auto-regressive}
\keyword{model}
\keyword{model,}
\keyword{multi-variate}

