% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{VAR_Gauss-class}
\alias{VAR_Gauss-class}
\title{VAR_Gauss}
\description{
A variable auto-regressive block which inherits from class \code{GMRF}. The primary difference is that this class is constructed using
temporal evolving functions. These functions are then used to construct a big precision matrix and mean vectors which are then passed on to the GMRF
constructor.
}
\details{

}

