% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{worldmap}
\alias{worldmap}
\title{World map}
\format{A data frame with 33971 rows and 6 variables:
\describe{
  \item{long}{longitude coordinate}
  \item{lat}{latitude coordinate}
  \item{group}{polygon (region) number}
  \item{order}{order of point in polygon boundary}
  \item{region}{region name}
  \item{subregion}{subregion name}
}}
\usage{
worldmap
}
\description{
This world map was extracted from the package \code{maps} v.3.0.1 by running \code{map_data("world")}. To reduce the data size, only every third point of this data frame is contained in \code{worldmap}.
}
\references{
Original S code by Becker, R.A. and Wilks, R.A. R version by Brownrigg, R. Enhancements by Minka, T.P. and Deckmyn, A. (2015) maps: Draw Geographical Maps, R package version 3.0.1.
}
\keyword{datasets}

