% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Am_data}
\alias{Am_data}
\title{Americium soil data}
\format{
A data frame with 212 rows and 4 variables:
\describe{
  \item{Easting}{Easting in metres}
  \item{Northing}{Northing in metres}
  \item{Am}{Americium concentration in 1000 counts per minute}
}
}
\usage{
Am_data
}
\description{
Between 1954 and 1963, nuclear devices were detonated at Area 13 of the Nevada Test Site in the United States, contaminating the surrounding soil with the radioactive element americium (Am). This data set contains Am concentrations in a spatial domain immediately surrounding Ground Zero (GZ)
}
\references{
Paul R, Cressie N (2011). “Lognormal block kriging for contaminated soil.” European Journal of Soil Science, 62, 337–345.
}
\keyword{datasets}
