% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basisfns.R
\name{auto_basis}
\alias{auto_basis}
\title{Automatic basis-function placement}
\usage{
auto_basis(m = plane(), data, regular = 1, nres = 2, prune = 0,
  subsamp = 10000, type = "Gaussian")
}
\description{
Automatic basis-function placement
}

