% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{auto_BAUs}
\alias{auto_BAUs}
\title{Auatomatic BAU generation}
\usage{
auto_BAUs(manifold, res = 2, cellsize = c(1, 1), type = "hex",
  data = NULL, convex = -0.05)
}
\description{
Autoatic BAU generation
}

