% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{concat}
\alias{concat}
\alias{concat,Basis-method}
\title{Concatenation}
\usage{
concat(...)

\S4method{concat}{Basis}(...)
}
\arguments{
\item{...}{a series of \code{MVST} objects}
}
\description{
Concatenates MVST objects of the same class together. This is primarily used to join up \code{GMRF_basis} blocks and \code{Obs} blocks together.
}
\examples{
data(icesat)
icesat_obs <- Obs(df=icesat)
icesat_2x <- concat(icesat_obs,icesat_obs)
}

