% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{distance}
\alias{distance}
\alias{distance,manifold-method}
\alias{distance,measure-method}
\title{Retrieve distance measure used}
\usage{
distance(d, x1, x2)

\S4method{distance}{measure}(d, x1, x2)

\S4method{distance}{manifold}(d, x1, x2)
}
\description{
Retrieve distance measure used
}

