% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/geometryfns.R
\docType{methods}
\name{distance}
\alias{distance}
\alias{distance,measure-method}
\alias{distance,manifold-method}
\title{Compute distance}
\usage{
distance(d, x1, x2 = NULL)

\S4method{distance}{measure}(d, x1, x2 = NULL)

\S4method{distance}{manifold}(d, x1, x2 = NULL)
}
\arguments{
\item{d}{object of class \code{measure} or \code{manifold}}

\item{x1}{first coordinate}

\item{x2}{second coordinate}
}
\description{
Compute distance using object of class \code{measure} or \code{manifold}.
}
\examples{
distance(sphere(),matrix(0,1,2),matrix(10,1,2))
distance(plane(),matrix(0,1,2),matrix(10,1,2))
}
\seealso{
\code{\link{real_line}}, \code{\link{plane}}, \code{\link{sphere}}, \code{\link{STplane}} and \code{\link{STsphere}} for constructing manifolds, and \code{\link{distances}} for the type of distances available.
}
