% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{distances}
\alias{Euclid_dist}
\alias{distances}
\alias{gc_dist}
\title{Pre-configured distances}
\usage{
Euclid_dist(dim = 2L)

gc_dist(R = NULL)
}
\arguments{
\item{dim}{dimension of Eucledian space}

\item{R}{great-circle radius}
}
\description{
Distance objects included in package
}
\details{
Initialises an object of class \code{measure} which contains a function \code{dist} used for computing the distance between two points.  Currently the Euclidean distance and the great-circle distance are included. Distances are computed using functions from the package \code{fields}.
}

