% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBmisc.R
\name{.concat_percentiles_to_df}
\alias{.concat_percentiles_to_df}
\title{Computation and concatenation of percentiles to a dataframe.}
\usage{
.concat_percentiles_to_df(data, X, name, percentiles, cred_mass)
}
\arguments{
\item{data}{The dataframe we will append percentiles to; the number of rows (if \code{X} is a matrix) or the length (if \code{X} is a list) of \code{X} must equal the number of rows in \code{data}}

\item{X}{a matrix (wherein rows correspond to prediction locations and columns are samples) of Monte Carlo samples}

\item{name}{The name of the quantity of interest. The names of the percentile columns are "name_percentile". In \code{FRK} it will be Y, mu, prob, or Z}

\item{percentiles}{a vector of scalars in [0, 100] specifying the desired percentiles; if \code{percentiles = NULL}, no percentiles are computed}

\item{cred_mass}{a scalar [0, 1] specifying the mass within the credible interval; if \code{cred_mass = NULL}, no HPD credible interval is computed}
}
\value{
The dataframe \code{data} with appended percentiles
}
\description{
Computes the percentiles or HPD interval bounds at each prediction location and appends 
the result to \code{data}. Note that we use percentiles rather than quantiles
because we including a "dot" (corresponding to the decimal place) in the 
dataframe column name may cause issues.
}
