% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBmisc.R
\name{.concat_percentiles_to_df}
\alias{.concat_percentiles_to_df}
\title{Computation and concatenation of percentiles to a dataframe.}
\usage{
.concat_percentiles_to_df(data, MC, percentiles)
}
\arguments{
\item{data}{The dataframe we will append percentiles to; the number of rows of the matrices in \code{MC} and  in \code{data} must be equal}

\item{MC}{List of matrices containing Monte Carlo samples}

\item{percentiles}{a vector of scalars in [0, 100] specifying the desired percentiles; if \code{percentiles = NULL}, no percentiles are computed}
}
\value{
The dataframe \code{data} with appended percentiles
}
\description{
Computes the percentiles or HPD interval bounds at each prediction location and appends 
the result to \code{data}. Note that we use percentiles rather than quantiles
because we including a "dot" (corresponding to the decimal place) in the 
dataframe column name may cause issues.
}
