% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.equal_within_tol}
\alias{.equal_within_tol}
\title{Test equality within a given tolerance.}
\usage{
.equal_within_tol(x, y, tol = 1e-08)
}
\arguments{
\item{x}{\code{R} object.}

\item{y}{\code{R} object we wish to compare to \code{x}.}
}
\value{
If \code{x} and \code{y} are single numbers, then the function
returns 1 if \code{x} and \code{y} are equal (within \code{tol}), and 0
otherwise. However matrices may also be passed, in which case the function
returns a matrix of equal size with elementwise comparisons.
}
\description{
Tests equality between objects \code{x} and \code{y} (within a given tolerance, \code{tol}).
The primary purpose of this function is to avoid deeming objects to be unequal if they differ
by some very tiny amount due to floating point inaccuracies.
Of particular note is that the function can accept a matrix argument for \code{x} and a single numeric
for \code{y}, and the output will be a matrix with elements 1 and 0 if elements of \code{x} are equal to
\code{y} or not, respectively; i.e., it does elementwise comparisons.
}
\examples{
.equal_within_tol(2, 2 + 0.00000000000001)
.equal_within_tol(2, 2 + 0.1)

A <- matrix(1:4, ncol = 2, byrow = TRUE)
B <- matrix(c(1:3, 5), ncol = 2, byrow = TRUE)
.equal_within_tol(A, B)
.equal_within_tol(A, 3)
}
