% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.FRKTMB_fit}
\alias{.FRKTMB_fit}
\title{Fitting stage of non-Gaussian FRK.}
\usage{
.FRKTMB_fit(M, optimiser, sigma2fs, ...)
}
\arguments{
\item{M}{An object of class \code{SRE}.
The \code{SRE} object slots particularly relevant to \code{FRKTMB_fit} are
\describe{
  \item{\code{K_type}}{A string indicating the desired formulation of the prior variance/precision matrix of eta.}
  \item{\code{response}}{A string indicating the assumed distribution of the response variable.}
  \item{\code{link}}{A string indicating the desired link function.}
  \item{\code{taper}}{A positve numeric indicating the strength of the covariance tapering (only applicable if \code{K_type == "covariance"}).}
}
Furthermore in the case of binomial or negative-binomial response variables,
the \code{data} slot of \code{M} must contain a column named \code{k}
which contains the 'known-constant' parameters for each observation
(the number of trials for binomial data, or the target number of successes for negative-binomial data).}

\item{optimiser}{the optimising function used for model fitting when \code{method = 'TMB'} (default is \code{nlminb}). Users may pass in a function object or a string corresponding to a named function. Optional parameters may be passed to \code{optimiser} via \code{...}. The only requirement of \code{optimiser} is that the first three arguments correspond to the initial parameters, the obejctive function, and the gradient, respectively (note that this may be achieved by rearranging the order of arguments before passing into \code{optimiser})}

\item{...}{other parameters passed on to \code{auto_basis} and \code{auto_BAUs} when calling \code{FRK}, or the user specified \code{optimiser} function when calling \code{FRK} or \code{SRE.fit}}
}
\value{
This function updates the following slots of \code{M}:
\describe{
  \item{Q_eta_xi}{An estimate of the joint precision matrix of all random effects in the model (the random weights \eqn{\eta} and fine-scale variation \eqn{\xi}).}
  \item{mu_eta}{Posterior expectation of the random weights \eqn{\eta}.}
  \item{mu_xi}{Posterior expectation of the fine-scale variation \eqn{\xi}.}
  \item{alphahat}{Estimate of the fixed-effects.}
  \item{sigma2fshat}{Estimate of the variance parameter of the fine-scale variation.}
  \item{phi}{Estimate of the dispersion parameter (only for applicable response distributions).}
  \item{log_likelihood}{The log-likelihood of the model evaluated at the final parameter estimates. Can be obtained by calling loglik(M).}
}
}
\description{
Performs model fitting using \code{TMB}.
Prepares an object of class \code{SRE} for the prediction stage
(performed internally with \code{\link{.FRKTMB_pred}}).
}
