% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.neighbour_matrix}
\alias{.neighbour_matrix}
\title{Neighbour matrix.}
\usage{
.neighbour_matrix(
  df,
  loc1 = "loc1",
  loc2 = "loc2",
  order = 1,
  diag_neighbours = FALSE
)
}
\arguments{
\item{df}{A dataframe containing spatial coordinates.}

\item{loc1}{A string indicating the name of the column storing the x-coordinate.}

\item{loc2}{A string indicating the name of the column storing the y-coordinate.}

\item{order}{If order == 1, only first order neighbours are considered. If order == 2, second order neighbours are also considered, and so on.}

\item{diag_neighbours}{Indicates whether to consider the diagonal neighbours. If FALSE (the default), only the horizontal and vertical neighbours are considered.}
}
\value{
A "neighbour" matrix with element (i, j), for i not equal to j,
equal to 1/l if basis functions i and j are lth order neighbours (provided \code{l <= order}), and 0 otherwise.
Diagonal elements indicate the row sums.
}
\description{
Creates a matrix \eqn{A} with elements \eqn{A_{i, j}} equal to 1 if basis
functions i and j (i not equal to j) are first order neighbours, 1/2 if they are second order neighbours, and so on.
Neighbours with a larger order than specified by \code{order} have 0 in this matrix.
The diagonal elements of \eqn{A} (i.e. \eqn{A_{i, i}}) indicate the row sums (if 
the \code{order == 1}, then it is the totalnumber of first order neighbours associated with basis function i).
}
\details{
This function is only designed for basis functions
at \emph{one resolution}. It also assumes the basis functions are in a
regularly-spaced lattice; the shape of the lattice is not important, however
there \emph{must} be constant spacing between basis functions in a given
direction (horizontal and vertical spacing can be different).
}
\seealso{
\code{\link{.sparse_Q}}
}
