% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{.plot_data}
\alias{.plot_data}
\title{Plot 2D spatial data}
\usage{
.plot_data(
  zdf,
  response,
  point_size = 1,
  lim = range(zdf[["z"]]),
  z = "z",
  x = "x",
  y = "y"
)
}
\arguments{
\item{zdf}{A \code{dataframe} containing spatial coordinates and the value of the response variable.}

\item{response}{A character indicating the assumed response distribution.}

\item{point_size}{Size of plotted points.}

\item{lim}{Controls the limits of the colour scale.}

\item{z}{A \code{string} indicating the data column.}

\item{x}{A \code{string} indicating the name of the column containing x-locations.}

\item{y}{A \code{string} indicating the name of the column containing the y-locations.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot 2D spatial data using a diverging red and blue palette if 
\code{response == "bernoulli"}, and a spectral palette otherwise.
}
\seealso{
\code{\link{.plot_map}}, \code{\link{.plot_all}}
}
