% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBfit.R
\name{.TMB_prep}
\alias{.TMB_prep}
\title{TMB data and parameter preparation.}
\usage{
.TMB_prep(M, est_finescale)
}
\arguments{
\item{M}{An object of class \code{SRE}.}
}
\value{
A list object containing:
\describe{
  \item{data}{The data.}
  \item{parameters}{The initialised parameters/fixed-effects/random-effects.}
}
}
\description{
Prepares data and parameter/fixed-effects/random-effects for input to \code{TMB}.
}
