% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{EmptyTheme}
\alias{EmptyTheme}
\title{Empty-plot theme}
\usage{
EmptyTheme()
}
\value{
Object of class \code{ggplot}
}
\description{
Formats a ggplot object for plotting with no annotations/grids.
}
\examples{
\dontrun{
X <- data.frame(x=runif(100),y = runif(100), z = runif(100))
EmptyTheme() + geom_point(data=X,aes(x,y,colour=z))
}
}
\keyword{ggplot}

