% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/basisfns.R
\docType{methods}
\name{eval_basis}
\alias{eval_basis}
\alias{eval_basis,Basis,SpatialPointsDataFrame-method}
\alias{eval_basis,Basis,SpatialPolygonsDataFrame-method}
\alias{eval_basis,Basis,matrix-method}
\alias{eval_basis,Basis-SpatialPointsDataFrame-method}
\alias{eval_basis,Basis-SpatialPolygonsDataFrame-method}
\alias{eval_basis,Basis-matrix-method}
\title{Evaluate basis functions at points or over polygons}
\usage{
eval_basis(basis, s, output = "list")

\S4method{eval_basis}{Basis,matrix}(basis, s, output = "matrix")

\S4method{eval_basis}{Basis,SpatialPointsDataFrame}(basis, s,
  output = "matrix")

\S4method{eval_basis}{Basis,SpatialPolygonsDataFrame}(basis, s,
  output = "matrix")
}
\description{
Evaluate basis functions at points or over polygons
}

