% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GRBFBasis-class}
\alias{GRBFBasis-class}
\title{Gaussian radial basis functions basis}
\description{
GRBFBasis inherits from the virtual class Basis. A GRBF basis functions is initialsied using the function \code{initGRBFbasis}.
}
\details{

}
\keyword{Basis}
\keyword{GRBF}
\keyword{functions,}

