% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{info_fit}
\alias{info_fit}
\alias{info_fit,SRE-method}
\title{Retrieve fit information for SRE model}
\usage{
info_fit(object)

\S4method{info_fit}{SRE}(object)
}
\arguments{
\item{object}{object of class \code{SRE}}
}
\description{
Takes an object of class \code{SRE} and returns a list containing all the relevant information on parameter estimation
}
\examples{
# See example in the help file for SRE
}
\seealso{
See \code{\link{SRE}} for more information on the SRE model and available fitting methods.
}
