% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\docType{methods}
\name{initialize,manifold-method}
\alias{initialize,manifold-method}
\title{manifold}
\usage{
\S4method{initialize}{manifold}(.Object)
}
\arguments{
\item{.Object}{\code{manifold} object passed up from lower-level constructor}
}
\description{
Manifold initialisation. This function should not be called directly as \code{manifold} is a virtual class.
}
