% Generated by roxygen2 (5.0.1): do not edit by hand
\docType{methods}
\name{initialize,manifold-method}
\alias{initialize,manifold-method}
\title{manifold}
\usage{
\S4method{initialize}{manifold}(.Object)
}
\arguments{
\item{.Object}{\code{manfold} object passed up from lower-level constructor}
}
\description{
Manifold initialisation. This function should not be called directly as \code{manifold} is a virtual class.
}

