% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{isea3h}
\alias{isea3h}
\title{ISEA Aperture 3 Hexagon (ISEA3H) Discrete Global Grid}
\format{A data frame with 766280 rows and 5 variables:
\describe{
  \item{id}{grid identification number within the given resolution}
  \item{lon}{longitude coordinate}
  \item{lat}{latitude coordinate}
  \item{res}{DGGRID resolution (0 -- 9)}
  \item{centroid}{A 0-1 variable, indicating whether the point describes the centroid of the polygon, or is a point describing the polygon itself}
}}
\usage{
isea3h
}
\description{
The data used here was obtained from http://webpages.sou.edu/~sahrk/dgg/isea.old/gen/isea3h.html and represent ISEA discrete global grids (DGGs) generated using the DGGRID software. The original .gen files were converted to a data frame using the function \code{dggrid_gen_to_df}, available with the \code{FRK} package.'
}
\references{
Sahr, K. (2008) Location coding on icosahedral aperture 3 hexagon discrete global grids. Computers, Environment and Urban Systems, 32(3):174-187.
}
\keyword{datasets}

