% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{loglik}
\alias{loglik}
\alias{loglik,SRE-method}
\title{Retrieve log-likelihood}
\usage{
loglik(object)

\S4method{loglik}{SRE}(object)
}
\arguments{
\item{object}{object of class \code{SRE}}
}
\description{
Takes an object of class \code{SRE} and returns the log-likelihood function.
}
\examples{
# See example in the help file for FRK
}
\seealso{
\code{\link{FRK}} for more information on how to construct and fit an SRE model, and other available methods for an object of class SRE.
}
