% Generated by roxygen2 (5.0.1): do not edit by hand
\docType{methods}
\name{manifold}
\alias{manifold}
\alias{manifold,Basis-method}
\title{Retrieve manifold}
\usage{
manifold(.Object)

\S4method{manifold}{Basis}(.Object)
}
\arguments{
\item{.Object}{\code{FRK} object}
}
\description{
Retrieve manifold from \code{FRK} object.
}
\examples{
G <-  radial_basis(manifold = plane(),
                   loc=matrix(0,1,2),
                   scale=0.2,
                   type="bisquare")
manifold(G)
}

