% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/geometryfns.R
\name{manifold}
\alias{manifold}
\alias{manifold,Basis-method}
\alias{manifold,TensorP_Basis-method}
\title{Retrieve manifold}
\usage{
manifold(.Object)

\S4method{manifold}{Basis}(.Object)

\S4method{manifold}{TensorP_Basis}(.Object)
}
\arguments{
\item{.Object}{\code{FRK} object}
}
\description{
Retrieve manifold from \code{FRK} object.
}
\examples{
G <-  local_basis(manifold = plane(),
                   loc=matrix(0,1,2),
                   scale=0.2,
                   type="bisquare")
manifold(G)
}
\seealso{
\code{\link{real_line}}, \code{\link{plane}}, \code{\link{sphere}}, \code{\link{STplane}} and \code{\link{STsphere}} for constructing manifolds.
}
