% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{NOAA_df_1990}
\alias{NOAA_df_1990}
\title{NOAA maximum temperature data for 1990--1993}
\format{
A data frame with 196,253 rows and 8 variables:
\describe{
  \item{year}{year of retrieval}
  \item{month}{month of retrieval}
  \item{day}{day of retrieval}
  \item{z}{dependent variable}
  \item{proc}{variable name (Tmax)}
  \item{id}{station id}
  \item{lon}{longitude coordinate of measurement station}
  \item{lat}{latitude coordinate of measurement station}
}
}
\usage{
NOAA_df_1990
}
\description{
Maximum temperature data obtained from the National Oceanic and Atmospheric
Administration (NOAA) for a part of the USA between 1990 and 1993 (inclusive).
See http://iridl.ldeo.columbia.edu/ SOURCES/.NOAA/.NCDC/.DAILY/.FSOD/.
}
\references{
National Climatic Data Center, March 1993: Local Climatological Data.
Environmental Information summary (C-2), NOAA-NCDC, Asheville, NC.
}
\keyword{datasets}
