% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{observed_BAUs}
\alias{observed_BAUs}
\alias{unobserved_BAUs}
\alias{observed_BAUs,SRE-method}
\alias{unobserved_BAUs,SRE-method}
\title{Observed (or unobserved) BAUs}
\usage{
observed_BAUs(object)

unobserved_BAUs(object)

\S4method{observed_BAUs}{SRE}(object)

\S4method{unobserved_BAUs}{SRE}(object)
}
\arguments{
\item{object}{object of class \code{SRE}}
}
\description{
Computes the indices (a numeric vector) of the observed (or unobserved) BAUs
}
\examples{
# See example in the help file for FRK
}
\seealso{
See \code{\link{FRK}} for more information on the SRE model and available fitting methods.
}
