% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/misc.R
\name{observed_BAUs}
\alias{observed_BAUs}
\alias{observed_BAUs,SRE-method}
\title{Indices of observed BAUs}
\usage{
observed_BAUs(object)

\S4method{observed_BAUs}{SRE}(object)
}
\arguments{
\item{object}{object of class \code{SRE}}
}
\description{
Computes the indices (a numeric vector) of the observed BAUs
}
\seealso{
\code{\link{unobserved_BAUs}}
}
