% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{opts_FRK}
\alias{opts_FRK}
\title{Options for FRK package}
\format{\preformatted{List of 2
 $ set:function (opt, value)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 4 11 7 5 11 5 4 7
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x975b3f0> 
 $ get:function (opt)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 8 11 11 5 11 5 8 11
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x975b3f0> 
}}
\usage{
opts_FRK
}
\description{
Options for FRK package
}
\keyword{datasets}

