% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{opts_FRK}
\alias{opts_FRK}
\title{FRK options}
\format{\preformatted{List of 2
 $ set:function (opt, value)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 5 16 8 5 16 5 5 8
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x41cead8> 
 $ get:function (opt)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 9 11 12 5 11 5 9 12
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x41cead8> 
}}
\usage{
opts_FRK
}
\description{
The main options list for the FRK package
}
\keyword{datasets}

