% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{plane}
\alias{plane}
\title{plane}
\usage{
plane(metric = Euclid_dist(dim = 2L))
}
\arguments{
\item{metric}{an object of class \code{measure}}
}
\description{
Plane initialisation.
}
\details{
A 2D plane is initialised using a \code{measure} object. By default, the measure object (\code{metric}) is the Euclidean distance in 2 dimensions. Distances are computed using functions from the \code{fields} package.
}
\examples{
P <- plane()
print(type(P))
print(sp::dimensions(P))
}

