% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,FEBasis,character-method}
\alias{plot,Obs,character-method}
\alias{plot,Obs_poly,character-method}
\title{Plot meshes and observations}
\usage{
plot(x, y, ...)

\S4method{plot}{FEBasis,character}(x, y, ...)

\S4method{plot}{Obs,character}(x, y, ...)

\S4method{plot}{Obs_poly,character}(x, y, ...)
}
\arguments{
\item{x}{an MVST object of class \code{Obs} or \code{Basis_GMRF}}

\item{y}{a character indicated which column to plot}

\item{...}{other parameter used to configure the plot. These include \code{max} (upperbound on colour scale) and \code{min} (lowerbound on colour scale)}
}
\value{
a ggplot2 object
}
\description{
This function is the general one used for plotting/visualising objects with MVST.
}
\examples{
\dontrun{
data(surf_fe)
Mesh <- initFEbasis(p=surf_fe$p, t = surf_fe$t, M = surf_fe$M, K = surf_fe$K)
Mesh["z"] <- sin(Mesh["x"]/1000)*cos(Mesh["y"]/1000)
g <- plot(Mesh,"z")

data(icesat)
icesat_obs <- Obs(df=icesat)
plot(subset(icesat_obs,t==2),"z",min=-0.5,max=0.5)}
}

