% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfns.R
\name{plot_spatial_or_ST}
\alias{plot_spatial_or_ST}
\title{Plot data from a Spatial*DataFrame or STFDF object}
\usage{
plot_spatial_or_ST(
  object,
  column_names,
  map_layer = NULL,
  subset_time = NULL,
  palette = "Spectral",
  ...
)
}
\arguments{
\item{object}{a Spatial*DataFrame or STFDF}

\item{column_names}{a vector of strings indicating the columns of the data to plot}

\item{map_layer}{(optional) a \code{ggplot} layer or object to add below the plotted layer, often a map}

\item{subset_time}{(optional) a vector of times to be included; applicable only for \code{STFDF} objects}

\item{palette}{the palette supplied to scale_*_distiller()}

\item{...}{optional arguments passed on to whatever geom is appropriate for the data (geom_point, geom_raster, or geom_polygon)}
}
\value{
a list of \code{ggplot} objects corresponding to the provided \code{column_names}. This list can then be supplied to, for example, \code{ggpubr::ggarrange()}
}
\description{
Plot data from a Spatial*DataFrame or STFDF object
}
\examples{
## See example in the help file for SRE
}
\seealso{
\code{\link{plot}}
}
