% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basisfns.R
\name{radial_basis}
\alias{radial_basis}
\title{Construct a set of radial basis functions}
\usage{
radial_basis(manifold = sphere(), loc = matrix(c(1, 0), nrow = 1),
  scale = 1, type = "Gaussian")
}
\description{
Construct a set of radial basis functions
}

