% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/SREutils.R
\name{remove_BAUs}
\alias{remove_BAUs}
\alias{remove_BAUs,SpatialPixelsDataFrame-method}
\alias{remove_BAUs,STFDF-method}
\title{Remove BAUs}
\usage{
remove_BAUs(BAUs, rmidx)

\S4method{remove_BAUs}{SpatialPixelsDataFrame}(BAUs, rmidx)

\S4method{remove_BAUs}{STFDF}(BAUs, rmidx)
}
\arguments{
\item{BAUs}{BAU object.}

\item{rmidx}{indices of basis functions to remove.}
}
\value{
\code{BAUs} with specified BAUs removed.
}
\description{
Takes a \code{BAU} object returns the \code{BAU} object with selected BAUs removed.
}
