% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{sphere}
\alias{sphere}
\title{sphere}
\usage{
sphere(radius = 6371)
}
\arguments{
\item{radius}{radius of sphere}
}
\description{
Initialisation of the S2 sphere.
}
\details{
A sphere is initialised using a \code{radius} parameter. By default, the radius \code{R} is equal to \code{R}=6371 km, the Earth's radius, while the measure used to compute distances on the sphere is the great-circle distance on a sphere of radius \code{R}.
}
\examples{
S <- sphere()
print(sp::dimensions(S))
}

