% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{sphere}
\alias{sphere}
\title{sphere}
\usage{
sphere(radius = 6371)
}
\arguments{
\item{radius}{radius of sphere}
}
\description{
Sphere initialisation.
}
\details{
A sphere is initialised using a \code{radius} parameter. By default, the radius \code{R} is equal to \code{R}=6371 km, the Earth's radius, while the measure used to compute distances on the sphere is the great-circle distance on a sphere of radius \code{R}. Currently the user does not have the option to alter the distance function used on the sphere. Distances are computed using functions from the \code{fields} package.
}
\examples{
S <- sphere()
print(sp::dimensions(S))
}

