% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{STplane}
\alias{STplane}
\title{plane in space-time}
\usage{
STplane(measure = Euclid_dist(dim = 3L))
}
\arguments{
\item{measure}{an object of class \code{measure}}
}
\description{
Initialisation of a 2D plane with a temporal dimension.
}
\details{
A 2D plane with a time component added is initialised using a \code{measure} object. By default, the measure object (\code{measure}) is the Euclidean distance in 3 dimensions, \link{Euclid_dist}.
}
\examples{
P <- STplane()
print(type(P))
print(sp::dimensions(P))
}

