% Generated by roxygen2 (5.0.1): do not edit by hand
\name{STsphere}
\alias{STsphere}
\title{Space-time sphere}
\usage{
STsphere(radius = 6371)
}
\arguments{
\item{radius}{radius of sphere}
}
\description{
Domain for modellng space-time fields over a sphere.
}
\details{
As with the spatial-only sphere, the sphere is initialised using a \code{radius} parameter. By default, the radius \code{R} is equal to \code{R}=6371 km, the Earth's radius, while the measure used to compute distances on the sphere is the great-circle distance on a sphere of radius \code{R}. Distances are computed using functions from the \code{fields} package.
}
\examples{
S <- STsphere()
print(sp::dimensions(S))
}

