% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{STsphere}
\alias{STsphere}
\title{Space-time sphere}
\usage{
STsphere(radius = 6371)
}
\arguments{
\item{radius}{radius of sphere}
}
\description{
Initialisation of an S2 sphere with a temporal dimensions
}
\details{
As with the spatial-only sphere, the sphere is initialised using a \code{radius} parameter. By default, the radius \code{R} is equal to \code{R}=6371 km, the Earth's radius, while the measure used to compute distances on the sphere is the great-circle distance on a sphere of radius \code{R}.
}
\examples{
S <- STsphere()
print(sp::dimensions(S))
}

