% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometryfns.R
\name{timeline}
\alias{timeline}
\title{timeline}
\usage{
timeline(metric = Euclid_dist(dim = 1L))
}
\arguments{
\item{metric}{an object of class \code{measure}}
}
\description{
Initialisation of a timeline (1D real line).
}
\details{
A time axis initialised using a \code{measure} object. By default, the measure object (\code{metric}) is the absolute differences.
}
\examples{
P <- timeline()
print(type(P))
print(sp::dimensions(P))
}

