% Generated by roxygen2 (5.0.1): do not edit by hand
\name{timeline}
\alias{timeline}
\title{timeline}
\usage{
timeline(metric = Euclid_dist(dim = 1L))
}
\arguments{
\item{metric}{an object of class \code{measure}}
}
\description{
Timeline initialisation.
}
\details{
A time axis initialised using a \code{measure} object. By default, the measure object (\code{metric}) is the absolute differences.
}
\examples{
P <- timeline()
print(type(P))
warning("CHECK: This may not work for ST plane")
print(sp::dimensions(P))
}

