% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/geometryfns.R
\docType{methods}
\name{type}
\alias{type}
\alias{type,manifold-method}
\title{Type of manifold}
\usage{
type(.Object)

\S4method{type}{manifold}(.Object)
}
\description{
Type of manifold
}

