% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/geometryfns.R
\docType{methods}
\name{type}
\alias{type}
\alias{type,manifold-method}
\title{Type of manifold}
\usage{
type(.Object)

\S4method{type}{manifold}(.Object)
}
\arguments{
\item{.Object}{object of class \code{Basis} or \code{manifold}}
}
\description{
Retrieve slot \code{type} from object
}
\examples{
S <- sphere()
print(type(S))
}
